% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjustDiagnosisIDs.R
\name{AdjustDiagnosisIDs}
\alias{AdjustDiagnosisIDs}
\title{AdjustDiagnosisIDs}
\usage{
AdjustDiagnosisIDs(Table, IDList_Duplicates, IDList_Associations)
}
\arguments{
\item{Table}{Table from partially curated CCP data set}
}
\value{
Data frame with adjusted DiagnosisID entries
}
\description{
After classification of duplicated and associated diagnosis entries, adjust DiagnosisID in related tables.
}
\author{
Bastian Reiter
}
