% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttritionTracker_Close.R
\name{AttritionTracker_Close}
\alias{AttritionTracker_Close}
\title{AttritionTracker_Close}
\usage{
AttritionTracker_Close(AttritionTracker)
}
\arguments{
\item{AttritionTracker}{A preformed tibble containing attrition information}
}
\value{
A tibble with finalized Attrition Tracker
}
\description{
AttritionTracker_Close
}
\author{
Bastian Reiter
}
