% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttritionTracker_Init.R
\name{AttritionTracker_Init}
\alias{AttritionTracker_Init}
\title{AttritionTracker_Init}
\usage{
AttritionTracker_Init()
}
\value{
A tibble
}
\description{
Initialize Attrition Tracker (auxiliary table to track sample sizes around data excluding operations)
}
\author{
Bastian Reiter
}
