% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AttritionTracker_Update.R
\name{AttritionTracker_Update}
\alias{AttritionTracker_Update}
\title{AttritionTracker_Update}
\usage{
AttritionTracker_Update(
  AttritionTracker,
  NewSampleSize,
  InclusionComment = "",
  ExclusionComment = ""
)
}
\arguments{
\item{AttritionTracker}{A preformed tibble}

\item{NewSampleSize}{Integer | Sample size at point of call in processing}

\item{InclusionComment}{String | Optional comment on reasons for inclusion}

\item{ExclusionComment}{String | Optional comment on reasons for exclusion}
}
\value{
A tibble (Attrition Tracker)
}
\description{
Adds entry to preformed Attrition Tracker
}
