% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentEventData.R
\name{AugmentEventData}
\alias{AugmentEventData}
\title{AugmentEventData}
\usage{
AugmentEventData(EventData, RuleCalls, ProgressBarObject = NULL)
}
\arguments{
\item{EventData}{\code{data.frame} - Usually ADS$Events}

\item{ProgressBarObject}{\code{progress::progress_bar} object - Optionally pass progress bar object to display progress}
}
\value{
\code{data.frame} EventData enhanced with new features
}
\description{
Auxiliary function within \code{\link{AugmentDataDS}}
}
\details{
Enhance event data entries with informative features implementing feature engineering rules defined in meta data
}
\author{
Bastian Reiter
}
