% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.DrawSampleDS.R
\name{CCP.DrawSampleDS}
\alias{CCP.DrawSampleDS}
\title{CCP.DrawSampleDS}
\usage{
CCP.DrawSampleDS(RawDataSetName.S = "CCP.RawDataSet", SampleSize.S = 100)
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'CCP.RawDataSet'}

\item{SampleSize.S}{\code{integer} - Number of patients in sample}
}
\value{
A \code{list} containing a subset of Raw Data Set
}
\description{
Draws a sample (subset) from Raw Data Set
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
