% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckDataSetDS.R
\name{CheckDataSetDS}
\alias{CheckDataSetDS}
\title{CheckDataSetDS}
\usage{
CheckDataSetDS(
  DataSetName.S = "RawDataSet",
  RequiredTableNames.S,
  RequiredFeatureNames.S
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{RequiredTableNames.S}{\code{character vector} - Names of tables that are expected/required to be in the data set - Default: Names of elements in list evaluated from \code{DataSetName.S}}

\item{RequiredFeatureNames.S}{\code{list} of \code{character vectors} - Features that are expected/required in each table of the data set - Default: Names of features in respective table}
}
\value{
A list containing information about Data Set tables
}
\description{
Checks existence and details of tables in a given DataSet (RDS / CDS / ADS) and returns an informative list object.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
