% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckRDSTablesDS.R
\name{CheckRDSTablesDS}
\alias{CheckRDSTablesDS}
\title{CheckRDSTablesDS}
\usage{
CheckRDSTablesDS(RawDataSetName.S = "RawDataSet")
}
\arguments{
\item{RawDataSetName.S}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list containing information about existence and completeness of RDS tables
}
\description{
Checks existence and completeness of RawDataSet (RDS) tables and returns an informative list object.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
