% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckTable.R
\name{CheckTable}
\alias{CheckTable}
\title{CheckTable}
\usage{
CheckTable(Table, RequiredFeatureNames = NULL)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble}}

\item{RequiredFeatureNames}{\code{character vector} - Optional names of required features - Default: \code{names(Table)}}
}
\value{
A \code{list} containing informative meta data about a \code{data.frame}
}
\description{
Checks out a data.frame and returns an informative list object.
}
\author{
Bastian Reiter
}
