% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisAssociation.R
\name{ClassifyDiagnosisAssociation}
\alias{ClassifyDiagnosisAssociation}
\title{ClassifyDiagnosisAssociation}
\usage{
ClassifyDiagnosisAssociation(
  DiagnosisEntries,
  RuleCalls,
  ProgressBarObject = NULL
)
}
\arguments{
\item{DiagnosisEntries}{Data frame (partially curated) that contains multiple diagnosis entries}

\item{ProgressBarObject}{Optionally pass object of type progress::progress_bar to display progress}
}
\value{
Data frame enhanced with classified diagnosis associations
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
In patients with multiple diagnosis entries: Identify and classify associations between diagnosis entries.
}
\author{
Bastian Reiter
}
