% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisAssociations.R,
%   R/ClassifyDiagnosisDuplicates.R
\name{ClassifyDiagnosisAssociations}
\alias{ClassifyDiagnosisAssociations}
\title{ClassifyDiagnosisAssociations}
\usage{
ClassifyDiagnosisAssociations(DiagnosisEntries, RulesProfile = "Default")

ClassifyDiagnosisAssociations(DiagnosisEntries, RulesProfile = "Default")
}
\arguments{
\item{DiagnosisEntries}{Data frame (partially curated) that contains multiple diagnosis entries}

\item{Method}{String | Determines the method being used for classification
\itemize{
\item "Rule-based" (Default)
\item "ML-Model" }}
}
\value{
Data frame enhanced with classified diagnosis associations

Data frame of classified duplicate diagnosis entries
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}

Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
In patients with multiple diagnosis entries: Identify and classify associations between diagnosis entries.

In patients with multiple diagnosis entries: Identify and classify duplicate diagnosis entries.
}
\author{
Bastian Reiter
}
