% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisEntries.R
\name{ClassifyDiagnosisEntries}
\alias{ClassifyDiagnosisEntries}
\title{ClassifyDiagnosisEntries}
\usage{
ClassifyDiagnosisEntries(DiagnosisEntries, Method = "Auto", Rules)
}
\arguments{
\item{Method}{String |}

\item{DataFrame}{Data frame (partially curated) that contains multiple diagnosis entries}
}
\value{
Data frame with classified diagnoses
}
\description{
In patients with multiple diagnosis entries: Distinguish pseudo-different from actually different diagnoses.
}
