% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisRedundancies.R
\name{ClassifyDiagnosisRedundancies}
\alias{ClassifyDiagnosisRedundancies}
\title{ClassifyDiagnosisRedundancies}
\usage{
ClassifyDiagnosisRedundancies(
  DiagnosisEntries,
  RulesProfile = "Default",
  ProgressBarObject = NULL
)
}
\arguments{
\item{DiagnosisEntries}{Data frame (partially curated) that (usually) contains multiple diagnosis entries}

\item{ProgressBarObject}{Optionally pass object of type progress::progress_bar to display progress}
}
\value{
Data frame of classified redundant diagnosis entries
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
In patients with multiple diagnosis entries: Identify and classify redundant diagnosis entries.
}
\author{
Bastian Reiter
}
