% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisRedundancy.R
\name{ClassifyDiagnosisRedundancy}
\alias{ClassifyDiagnosisRedundancy}
\title{ClassifyDiagnosisRedundancy}
\usage{
ClassifyDiagnosisRedundancy(
  DiagnosisEntries,
  RuleCalls,
  ProgressBarObject = NULL
)
}
\arguments{
\item{DiagnosisEntries}{Data frame (partially curated) that (usually) contains multiple diagnosis entries}

\item{ProgressBarObject}{Optionally pass object of type progress::progress_bar to display progress}
}
\value{
Data frame of consolidated diagnosis entries
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
In patients with multiple diagnosis entries: Identify, classify and remove redundant diagnosis entries.
}
\author{
Bastian Reiter
}
