% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTable.R
\name{CleanTable}
\alias{CleanTable}
\title{CleanTable}
\usage{
CleanTable(
  Table,
  TableNameLookup,
  RemoveRedundantEntries = TRUE,
  FeatureObligations
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble} - The table object to be cleaned}

\item{TableNameLookup}{\code{character} - The table name(s) to be looked up in 'FeatureObligations$RuleSet'. Can be a single string or a character vector.}

\item{RemoveRedundantEntries}{\code{logical} - Whether redundant entries should be removed}

\item{FeatureObligations}{\code{list}
\itemize{\item RuleSet \code{data.frame}
\item RuleSet.Profile \code{character} - Profile name defining strict and trans-feature rules for obligatory feature content. Profile name must be stated in \code{FeatureObligations$RuleSet}}}
}
\value{
\code{tibble} - Clean table
}
\description{
Auxiliary function in \code{dsCCPhos::CurateDataDS()} to perform exclusion of invalid table entries.
}
\author{
Bastian Reiter
}
