% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileClassificationCall.R
\name{CompileClassificationCall}
\alias{CompileClassificationCall}
\title{CompileClassificationCall}
\usage{
CompileClassificationCall(
  TargetFeature,
  PredictorFeatures,
  RuleSet,
  RuleProfile,
  ValueIfNoRuleMet = NA
)
}
\arguments{
\item{TargetFeature}{String | Name of feature that is being engineered}

\item{PredictorFeatures}{Vector | Character vector containing the names of all determining / predictor features}

\item{RuleSet}{Data frame | Contains predefined set of rules}

\item{RuleProfile}{String | Profile name stated in rule set data frame}

\item{ValueIfNoRuleMet}{String | Value (as string) of TargetFeature if no rules are fulfilled | Default = "NA"}
}
\value{
String containing an unevaluated dplyr::case_when() expression
}
\description{
Auxiliary function within \code{\link{ClassifyDiagnosisAssociations}}
}
\details{
Based on a predefined rule set data frame, compile a dplyr::case_when() expression and return it as a string
}
\author{
Bastian Reiter
}
