% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileHarmonizationRules.R
\name{CompileHarmonizationRules}
\alias{CompileHarmonizationRules}
\title{CompileHarmonizationRules}
\usage{
CompileHarmonizationRules(TableName, RuleSet, RuleSet.Profile)
}
\arguments{
\item{TableName}{\code{string} - Name of table that contains features to be transformed}

\item{RuleSet}{\code{data.frame} - Contains predefined set of harmonization rules}

\item{RuleSet.Profile}{\code{string} - Profile name stated in rule set data frame}
}
\value{
\code{string} containing arguments for \code{dplyr::mutate()}
}
\description{
Auxiliary function within \code{\link{CurateDataDS}} or \code{\link{HarmonizeData}} respectively
}
\details{
Based on a predefined rule set data frame, compile argument expressions for dplyr::mutate()
}
\author{
Bastian Reiter
}
