% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileHashTables.R
\name{CompileHashTables}
\alias{CompileHashTables}
\title{CompileHashTables}
\usage{
CompileHashTables(TableName, RuleSet, RuleSet.Profile)
}
\arguments{
\item{TableName}{\code{string} - Name of table that contains features to be transformed}

\item{RuleSet}{\code{data.frame} - Contains predefined set of transformation rules}

\item{RuleSet.Profile}{\code{string} - Profile name stated in 'RuleSet'}
}
\value{
\code{list} of named vectors (serving as hash tables)
}
\description{
Auxiliary function within \code{\link{CurateDataDS}} or \code{\link{HarmonizeData}} respectively
}
\details{
Based on a predefined rule set data frame, compile hash tables for direct value transformation
}
\author{
Bastian Reiter
}
