% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTransFeatureRule.R
\name{CompileTransFeatureRules}
\alias{CompileTransFeatureRules}
\title{CompileTransFeatureRules}
\usage{
CompileTransFeatureRules(PseudoCodeRules)
}
\arguments{
\item{PseudoCodeRules}{\code{character vector} containing pseudo-code rules about feature obligations}
}
\value{
\code{character vector} containing unevaluated expressions to be used in dplyr::mutate() calls
}
\description{
Auxiliary function to compile unevaluated R expression from trans-feature obligation rules stated as pseudo-code
}
\details{
Based on a predefined rule set data frame, compile a dplyr::case_when() expression and return it as a string
}
\author{
Bastian Reiter
}
