% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTransformationRules.R
\name{CompileTransformationRules}
\alias{CompileTransformationRules}
\title{CompileTransformationRules}
\usage{
CompileTransformationRules(TableName, RuleSet, RuleProfile)
}
\arguments{
\item{TableName}{String | Name of table that contains features to be transformed}

\item{RuleSet}{Data frame | Contains predefined set of transformation rules}

\item{RuleProfile}{String | Profile name stated in rule set data frame}
}
\value{
String containing arguments for dplyr::mutate()
}
\description{
Auxiliary function within \code{\link{CurateDataDS}} or \code{\link{TransformData}} respectively
}
\details{
Based on a predefined rule set data frame, compile argument expressions for dplyr::mutate()
}
\author{
Bastian Reiter
}
