% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileTransformativeAssignments.R
\name{CompileTransformativeAssignments}
\alias{CompileTransformativeAssignments}
\title{CompileTransformativeAssignments}
\usage{
CompileTransformativeAssignments(
  TransformativeExpressions,
  TransformativeExpressions.Profile,
  FeatureNames
)
}
\arguments{
\item{TransformativeExpressions}{\code{data.frame} - Contains expressions that transform data values}

\item{TransformativeExpressions.Profile}{\code{string} - Profile selected in \emph{TransformativeExpressions}}

\item{FeatureNames}{\code{character vector} - Names of features to be transformed}
}
\value{
\code{string} containing assignments for \code{dplyr::mutate()}
}
\description{
Auxiliary function within \code{\link{CurateDataDS}} or \code{\link{HarmonizeData}} respectively
}
\details{
Based on a given meta data.frame, compile assignment expressions for dplyr::mutate()
}
\author{
Bastian Reiter
}
