% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectToTestSevers.R
\name{ConnectToTestServers}
\alias{ConnectToTestServers}
\title{ConnectToTestServers}
\usage{
ConnectToTestServers()
}
\value{
A list of DSConnection objects
}
\description{
Establishes connections to two CCP test bridgeheads.
}
