% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCheckpointsDS.R
\name{CreateCheckpointsDS}
\alias{CreateCheckpointsDS}
\title{CreateCheckpointsDS}
\usage{
CreateCheckpointsDS()
}
\value{
A named vector
}
\description{
Creates named vector in server R session containing status data on defined processing checkpoints
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
