% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateEventFeatures.R
\name{CreateEventFeatures}
\alias{CreateEventFeatures}
\title{CreateEventFeatures}
\usage{
CreateEventFeatures(EventData, RuleSet, Profile, ProgressBarObject = NULL)
}
\arguments{
\item{EventData}{\code{data.frame} - Usually ADS$Events}

\item{RuleSet}{\code{data.frame} - Set of rules for feature engineering - Default: \code{\link{dsCCPhos::Meta_EventFeatures}}}

\item{Profile}{\code{character} - Profile name defining rule set to be used for event feature engineering. Profile name must be stated in \code{RuleSet} - Default: 'Default'}

\item{ProgressBarObject}{\code{progress::progress_bar} object - Optionally pass progress bar object to display progress}
}
\value{
\code{data.frame} EventData enhanced with new features
}
\description{
Auxiliary function within \code{\link{AugmentDataDS}}
}
\details{
Enhance event data entries with informative features implementing feature engineering rules defined in meta data
}
\author{
Bastian Reiter
}
