% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurationReportDS.R
\name{CurationReportDS}
\alias{CurationReportDS}
\title{CurationReportDS}
\usage{
CurationReportDS(Name_CurationOutput = "CurationOutput")
}
\arguments{
\item{Name_CurationOutput}{String | Name of object on server previously assigned by dsCCPhos::CurateData() | Default: 'CurationOutput'}
}
\value{
A list of data frames containing information about occurring data values in different Curation stages ("Raw" / "Transformed" / "Final")
}
\description{
What it does
}
\details{
Server-side AGGREGATE method
}
