% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataListToDataModel.R
\name{DataListToDataModel}
\alias{DataListToDataModel}
\title{DataListToDataModel}
\usage{
DataListToDataModel(DataList)
}
\arguments{
\item{DataList}{}
}
\value{
A list of data frames describing a data model
}
\description{
Take a list of data frames and return the underlying data model
}
