% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModelToCSV.R
\name{DataModelToCSV}
\alias{DataModelToCSV}
\title{DataModelToCSV}
\usage{
DataModelToCSV(DataModel, OutputPath = "./DataModel.csv")
}
\arguments{
\item{DataModel}{A list of data frames describing a data model}

\item{OutputPath}{A string describing the file path of the created .csv-file}
}
\value{
No return, creates a .csv-file.
}
\description{
DataModelToCSV
}
\author{
Bastian Reiter
}
