% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataModelToDB.R
\name{DataModelToDB}
\alias{DataModelToDB}
\title{DataModelToDB}
\usage{
DataModelToDB(DataModel, DBPath = ":memory:")
}
\arguments{
\item{DataModel}{A list of data frames describing a data model}

\item{DBPath}{}
}
\value{
A data base connection object
}
\description{
Make (virtual) data base from a data model
}
\author{
Bastian Reiter
}
