% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrawSampleDS.R
\name{DrawSampleDS}
\alias{DrawSampleDS}
\title{DrawSampleDS}
\usage{
DrawSampleDS(RawDataSetName.S = "RawDataSet", SampleSize.S = "100")
}
\arguments{
\item{RawDataSetName.S}{\code{string} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{SampleSize}{\code{string} - Number of patients in sample}
}
\value{
A \code{list} containing a subset of Raw Data Set
}
\description{
Draws a sample (subset) from Raw Data Set
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
