% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsCCPhos-package.R
\docType{package}
\name{dsCCPhos-package}
\alias{dsCCPhos}
\alias{dsCCPhos-package}
\title{dsCCPhos: Enabling Data Science with the DKTK-CCP}
\description{
Server-side DataSHIELD R package to enable Data Science with the German Cancer Consortium's Clinical Communication Platform (CCP)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ccphos.netlify.app/}
  \item \url{https://github.com/BastianReiter/dsCCPhos}
}

}
\author{
\strong{Maintainer}: Bastian Reiter \email{b.reiter@med.uni-frankfurt.de} (\href{https://orcid.org/0000-0002-1088-0502}{ORCID})

}
\keyword{internal}
