% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcludeCuratedDataDS.R
\name{ExcludeCuratedDataDS}
\alias{ExcludeCuratedDataDS}
\title{ExcludeCuratedDataDS}
\usage{
ExcludeCuratedDataDS(Name_RawDataSet = "RawDataSet")
}
\arguments{
\item{Name_CurationOutput}{String | Name of the list object created by CurateDataDS() | Default: 'CurationOutput'}
}
\value{
The Curation Output with modified Curated Data Set
}
\description{
Exclude data from Curated Data Set
}
\details{
Server-side ASSIGN method
}
