% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractFromListDS.R
\name{ExtractFromListDS}
\alias{ExtractFromListDS}
\title{ExtractFromListDS}
\usage{
ExtractFromListDS(ListName.S, ObjectName.S)
}
\arguments{
\item{ListName.S}{String | Name of a list object on server}

\item{ObjectName.S}{String | Name of object inside list}
}
\value{
The object to be extracted
}
\description{
Auxiliary function to extract objects from list on server
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
