% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterTable.R
\name{FilterTableDS}
\alias{FilterTableDS}
\title{FilterTableDS}
\usage{
FilterTableDS(TableName, FilterExpressions)
}
\arguments{
\item{FilterExpressions}{}

\item{TableName.S}{\code{string} - Name of Table A on server}
}
\value{
A \code{data.frame} resulting from filter operation
}
\description{
Filter a table on server, making use of \code{dplyr} and \code{stringr}.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
