% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalizeDataTransformation.R
\name{FinalizeDataTransformation}
\alias{FinalizeDataTransformation}
\title{FinalizeDataTransformation}
\usage{
FinalizeDataTransformation(
  TargetVector,
  TableName,
  FeatureName,
  ExcludeIneligibleValues = TRUE,
  ConvertToFactor = FALSE,
  AssignFactorLabels = FALSE
)
}
\arguments{
\item{TargetVector}{Vector | Feature to be finalized}

\item{TableName}{String | For look-up purposes: Name of parenting table within Curated Data Set (CDS)}

\item{FeatureName}{String | For look-up purposes: Name of the feature to be finalized}

\item{ExcludeIneligibleValues}{Logical | Whether to set all ineligible values NA | Default: TRUE}

\item{ConvertToFactor}{Logical | Whether to convert vector to factor | Default: FALSE}

\item{AssignFactorLabels}{Logical | Whether to assign factor labels during factor conversion | Default: FALSE}
}
\value{
A vector (or factor)
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
Ineligible data (including data that could not be transformed) is optionally turned into NA.
Optional factor conversion establishes level order where appropriate.
Value eligibility and factor information is defined in dsCCPhos::Meta_ValueSets.
}
\author{
Bastian Reiter
}
