% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindRedundantEntries.R
\name{FindRedundantEntries}
\alias{FindRedundantEntries}
\title{FindRedundantEntries}
\usage{
FindRedundantEntries(
  Table,
  PrimaryKeyFeature,
  DiscriminatoryFeatures,
  EssentialFeatures,
  RemoveRedundantEntries = FALSE
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble}}

\item{PrimaryKeyFeature}{\code{character} - Name of primary key feature}

\item{DiscriminatoryFeatures}{\code{vector} - Names of features that are used to strictly discriminate between different table entries (used in \code{group_by}-statement)}

\item{EssentialFeatures}{\code{vector} - Names of features that are considered essential in the informational value of a table entry}

\item{RemoveRedundantEntries}{\code{logical} - Indicates whether to remove table entries that are considered redundant. If set to \code{FALSE}, the redundant entries are marked as such and preserved. - Default: \code{FALSE}}
}
\value{
\code{data.frame} or \code{tibble} cleaned of redundant entries
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
Identify and remove any entries that contain less essential information compared to a previous entry.
}
\author{
Bastian Reiter
}
