% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FormatData.R
\name{FormatData}
\alias{FormatData}
\title{FormatData}
\usage{
FormatData(TargetVector, Type)
}
\arguments{
\item{TargetVector}{Vector that formatting is performed on}

\item{Type}{\code{character} - Character string defining data type other than character}
}
\value{
A vector
}
\description{
Format data based on feature type specifications defined in meta data
}
\author{
Bastian Reiter
}
