% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetClassificationCall.R
\name{GetClassificationCall}
\alias{GetClassificationCall}
\title{GetClassificationCall}
\usage{
GetClassificationCall(
  TargetFeature,
  DeterminingFeatures,
  RuleSet,
  ValueIfNoRuleMet = "NA"
)
}
\arguments{
\item{TargetFeature}{String | Name of feature that is being engineered}

\item{DeterminingFeatures}{Vector | Character vector containing the names of all determining features}

\item{RuleSet}{Data frame | Contains predefined set of rules}

\item{ValueIfNoRuleMet}{String | Value (as string) of TargetFeature if no rules are fulfilled | Default = "NA"}
}
\value{
String containing an unevaluated R expression
}
\description{
Auxiliary function within \code{\link{ClassifyDiagnosisAssociations}}
}
\details{
Based on a predefined rule set data frame, construct an R expression and return it as a string
}
\author{
Bastian Reiter
}
