% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCrossTabDS.R
\name{GetCrossTabDS}
\alias{GetCrossTabDS}
\title{GetCrossTabDS}
\usage{
GetCrossTabDS(TableName.S, FeatureNames.S)
}
\arguments{
\item{TableName.S}{\code{string} - Name of \code{data.frame}}

\item{FeatureNames.S}{\code{string} - Names of features separated by ','}
}
\value{
A \code{list} containing
\itemize{ \item CrossTab (\code{data.frame})
\item ChiSq.PValue (\code{numeric}) }
}
\description{
Create cross tab (contingency table) with arbitrary number of table features.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
