% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCurationReportDS.R
\name{GetCurationReportDS}
\alias{GetCurationReportDS}
\title{GetCurationReportDS}
\usage{
GetCurationReportDS(Name_CurationOutput = "CurationOutput")
}
\arguments{
\item{Name_CurationOutput}{String | Name of object on server previously assigned by dsCCPhos::CurateData() | Default: 'CurationOutput'}
}
\value{
A list of data frames containing information about occurring data values in different Curation stages ("Raw" / "Transformed" / "Final")
}
\description{
What it does
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
