% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataSetCheckDS.R
\name{GetDataSetCheckDS}
\alias{GetDataSetCheckDS}
\title{GetDataSetCheckDS}
\usage{
GetDataSetCheckDS(
  DataSetName.S,
  RequiredTableNames.S = NULL,
  RequiredFeatureNames.S = NULL,
  AssumeCCPDataSet.S = FALSE
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{RequiredTableNames.S}{\code{character vector} - Names of tables that are expected/required to be in the data set - Default: Names of elements in list evaluated from \code{DataSetName.S}}

\item{RequiredFeatureNames.S}{\code{list} of \code{character vectors} - Features that are expected/required in each table of the data set - Default: Names of features in respective table}

\item{AssumeCCPDataSet.S}{\code{logical} - Whether or not the data set to be checked out is one of the main data sets used in CCPhos - Default: FALSE}
}
\value{
A \code{list} containing meta data about tables in a data set
}
\description{
Checks out a given data set (list of data.frames) and returns an informative list object.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
