% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDFInfo.R
\name{GetDFInfo}
\alias{GetDFInfo}
\title{GetDFInfo}
\usage{
GetDFInfo(
  DataFrame,
  IncludeRandomExampleValues = FALSE,
  IncludeRandomRow = FALSE
)
}
\arguments{
\item{DataFrame}{A data frame or tibble}

\item{IncludeRandomRow}{}
}
\value{
A tibble containing information about a data frame's features
}
\description{
Returns information about a data frame's features and data types
}
