% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFeatureInfoDS.R
\name{GetFeatureInfoDS}
\alias{GetFeatureInfoDS}
\title{GetFeatureInfoDS}
\usage{
GetFeatureInfoDS(TableName.S, FeatureName.S)
}
\arguments{
\item{TableName.S}{\code{string} - Name of the data frame that contains the feature}

\item{FeatureName.S}{\code{string} - Name of feature}
}
\value{
A \code{list}
}
\description{
Obtain data about feature type and sample size.
}
\author{
Bastian Reiter
}
