% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFrequencyTableDS.R
\name{GetFrequencyTableDS}
\alias{GetFrequencyTableDS}
\title{GetFrequencyTableDS}
\usage{
GetFrequencyTableDS(TableName.S, FeatureName.S, GroupingFeatureName.S = NULL)
}
\arguments{
\item{TableName.S}{\code{string} | Name of the Data frame that contains the feature}

\item{FeatureName.S}{\code{string} | Name of feature}

\item{GroupingFeatureName.S}{\code{string} | Name of optional grouping feature}
}
\value{
A \code{tibble} containing absolute and relative frequencies
}
\description{
Return table of absolute and relative value frequencies for a nominal / ordinal feature.
}
\author{
Bastian Reiter
}
