% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFuzzyStringMatches.R
\name{GetFuzzyStringMatches}
\alias{GetFuzzyStringMatches}
\title{GetFuzzyStringMatches}
\usage{
GetFuzzyStringMatches(
  Vector,
  EligibleStrings,
  PreferredMethod = "jw",
  FindBestMethod = FALSE,
  Tolerance = 0.3,
  Preprocessing.FlattenCase = TRUE,
  Preprocessing.RemoveAllWhiteSpace = FALSE,
  Preprocessing.SquishWhiteSpace = TRUE,
  StringdistArguments = list()
)
}
\arguments{
\item{Vector}{\code{character vector} - Containing original values}

\item{EligibleStrings}{\code{character vector} - A set of valid/eligible strings}

\item{PreferredMethod}{\code{string} - Selects the method \code{stringdist()} uses (see \code{stringdist} documentation) - Default: 'jw'}

\item{FindBestMethod}{\code{logical} - Indicating whether this function should try out all available methods and choose the one that yields the best result in terms of highest proportion of correctly matched vector elements. - Default: \code{FALSE}}

\item{Tolerance}{\code{double} - Number between 0 and 1 relating to normalized distance between to strings (1 meaning furthest distance, 0 meaning no distance). If a string in 'Vector' is not similar enough to any of the 'EligibleStrings' and its minimal harmonized distance exceeds this number, it is set \code{NA}. Default: 0.3}

\item{Preprocessing.FlattenCase}{\code{logical} - If set to \code{TRUE} all letters in both the input and valid strings are transformed to lower case}

\item{Preprocessing.RemoveAllWhiteSpace}{\code{logical} - If set to \code{TRUE} all white space in both the input and valid strings is being removed}

\item{Preprocessing.SquishWhiteSpace}{\code{logical} - If set to \code{TRUE} all leading and trailing white space in both the input and valid strings is being removed and internal white space removed to a single space}

\item{StringdistArguments}{\code{list} containing additional arguments used in \code{stringdist::stringdist()}}
}
\value{
\code{character vector} of transformed strings
}
\description{
Employ fuzzy string matching (using package \code{stringdist}) to try and match elements of a character vector to a set of valid strings and return the transformed vector.
It is designed to maximize effectiveness and computational efficiency.
}
\author{
Bastian Reiter
}
