% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObjectInfoDS.R
\name{GetObjectInfoDS}
\alias{GetObjectInfoDS}
\title{GetObjectInfoDS}
\usage{
GetObjectInfoDS(ObjectName.S)
}
\arguments{
\item{ObjectName.S}{String | Name of object on server}
}
\value{
A list containing information about existence and class of object
}
\description{
Similar to dsBase::testObjExistsDS(). Tests whether a certain object exists physically on server and provides meta info about it.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
