% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetObjectMetaDataDS.R
\name{GetObjectMetaDataDS}
\alias{GetObjectMetaDataDS}
\title{GetObjectMetaDataDS}
\usage{
GetObjectMetaDataDS(ObjectName.S)
}
\arguments{
\item{ObjectName.S}{\code{string} - Name of object on server}
}
\value{
A \code{list} containing meta data
}
\description{
Collect meta data about an R object and return it as a list.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
