% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetOutcomeMeasuresDS.R
\name{GetOutcomeMeasuresDS}
\alias{GetOutcomeMeasuresDS}
\title{GetOutcomeMeasuresDS}
\usage{
GetOutcomeMeasuresDS(Name_AugmentationOutput = "AugmentationOutput")
}
\arguments{
\item{Name_AugmentationOutput}{String | Name of the list object created by AugmentDataDS() | Default: 'AugmentationOutput'}
}
\value{
ABC
}
\description{
Calculate clinical outcome measures (COM) from Augmented Data Set (ADS)
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
