% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetRDSValidationReportDS.R
\name{GetRDSValidationReportDS}
\alias{GetRDSValidationReportDS}
\title{GetRDSValidationReportDS}
\usage{
GetRDSValidationReportDS(RawDataSetName.S = "RawDataSet")
}
\arguments{
\item{RawDataSetName.S}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list containing validation reports
}
\description{
Performs validation operations on Raw Data Set (RDS) and returns a list containing informative data frames.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
