% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetReportingObjectDS.R
\name{GetReportingObjectDS}
\alias{GetReportingObjectDS}
\title{GetReportingObjectDS}
\usage{
GetReportingObjectDS(ObjectName.S)
}
\arguments{
\item{ObjectName.S}{\code{string} - Name of reporting object on server}
}
\value{
The reporting object
}
\description{
Transports a reporting object from server to client. Its name must be on a list of permitted object names to ensure data privacy.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
