% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleStatisticsDS.R
\name{GetSampleStatisticsDS}
\alias{GetSampleStatisticsDS}
\title{GetSampleStatisticsDS}
\usage{
GetSampleStatisticsDS(
  TableName,
  MetricFeatureName,
  GroupingFeatureName = NULL,
  na.rm = FALSE
)
}
\arguments{
\item{TableName}{String | Name of the Data frame that contains the metric feature}

\item{MetricFeatureName}{String | Name of metric feature}

\item{GroupingFeatureName}{String | Name of optional grouping feature}

\item{na.rm}{Logical | Whether NA values should be removed before calculating statistics}
}
\value{
A tibble containing the following statistics: sample size (N), minimum, maximum, Q1, Median, Q3, MAD, mean, standard deviation
}
\description{
Calculate common sample statistics for a feature of a data frame
}
