% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSurvModelDS.R
\name{GetSurvModelDS}
\alias{GetSurvModelDS}
\title{GetSurvModelDS}
\usage{
GetSurvModelDS(TableName.S, TimeFeature.S, EventFeature.S, Covariates.S = NULL)
}
\arguments{
\item{TableName.S}{\code{string} | Name of the data frame that holds time and event features}

\item{TimeFeature.S}{\code{string} | Name of time feature}

\item{EventFeature.S}{\code{string} | Name of event feature}

\item{Covariates.S}{\code{character vector}}
}
\value{
A \code{survival curve object}
}
\description{
Return Kaplan-Meier curve
}
