% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTableCheckDS.R
\name{GetTableCheckDS}
\alias{GetTableCheckDS}
\title{GetTableCheckDS}
\usage{
GetTableCheckDS(
  TableName.S,
  RequiredFeatureNames.S = NULL,
  GetTemplate.S = FALSE
)
}
\arguments{
\item{TableName.S}{\code{string} - Name of \code{data.frame} or \code{tibble}}

\item{RequiredFeatureNames.S}{\code{character vector} - Optional names of required features - Default: \code{names()} applied to Table evaluated from \code{TableName.S}}

\item{GetTemplate.S}{\code{logical} - If set to \code{TRUE}, the function returns a template incorporating required feature names without actually checking an existing table}
}
\value{
A list containing informative meta data about a \code{data.frame}
}
\description{
Wrapper function for \code{dsCCPhos::CheckTable()}. Checks out a data.frame and returns an informative list object.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
