% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetTTEModelDS.R
\name{GetTTEModelDS}
\alias{GetTTEModelDS}
\title{GetTTEModelDS}
\usage{
GetTTEModelDS(
  TableName.S,
  TimeFeature.S,
  EventFeature.S,
  ModelType.S = "survfit",
  CovariateA.S = NULL,
  CovariateB.S = NULL,
  CovariateC.S = NULL,
  MinFollowUpTime.S = 1
)
}
\arguments{
\item{TableName.S}{\code{string} - Name of the data frame that holds time and event features}

\item{TimeFeature.S}{\code{string} - Name of time feature}

\item{EventFeature.S}{\code{string} - Name of event feature}

\item{ModelType.S}{\code{string} - Function name of different TTE models implemented in \code{survival} package:
\itemize{\item 'survfit'
\item 'survdiff'
\item 'coxph'}}

\item{CovariateA.S}{\code{string} - Name of optional Covariate A}

\item{CovariateB.S}{\code{string} - Name of optional Covariate B}

\item{CovariateC.S}{\code{string} - Name of optional Covariate C}

\item{MinFollowUpTime.S}{\code{integer} - Optional minimum of observed follow up time}
}
\value{
A Time-to-Event model object
}
\description{
Return Time-to-Event model implemented in the \code{survival} package
}
