% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetValidationReportDS_RawData.R
\name{GetValidationReportDS_RawData}
\alias{GetValidationReportDS_RawData}
\title{GetValidationReportDS_RawData}
\usage{
GetValidationReportDS_RawData(Name_RawDataSet = "RawDataSet")
}
\arguments{
\item{Name_RawDataSet}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list containing the validation report
}
\description{
Performs validation operations on Raw Data Set (RDS) and returns a report.
}
\details{
Server-side AGGREGATE method
}
