% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeData.R
\name{HarmonizeData}
\alias{HarmonizeData}
\title{HarmonizeData}
\usage{
HarmonizeData(DataFrame, TableName, Methods = NULL, RuleSet, RuleSet.Profile)
}
\arguments{
\item{DataFrame}{\code{data.frame} containing data to be transformed}

\item{TableName}{\code{string} - Name of the table (to enable mapping to harmonization rules)}

\item{Methods}{\code{data.frame}}

\item{RuleSet}{\code{data.frame} - Contains predefined set of harmonization rules}

\item{RuleSet.Profile}{\code{string} - Profile name stated in 'RuleSet'}
}
\value{
The input \code{data.frame} with transformed data values
}
\description{
Harmonizes data based on a given rule set. First, general harmonizing transformations are performed using regular expression based rules. Second, for direct replacement of values hash tables are being compiled and applied.
}
\details{
Uses dplyr::mutate() to transform data based on harmonization rules defined in a given rule set (Default: dsCCPhos::RuleSet_RawDataHarmonization)
}
\author{
Bastian Reiter
}
