% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinTablesDS.R
\name{JoinTablesDS}
\alias{JoinTablesDS}
\title{JoinTablesDS}
\usage{
JoinTablesDS(
  TableNameA.S,
  TableNameB.S,
  ByStatement.S,
  JoinType.S = "left_join"
)
}
\arguments{
\item{TableNameA.S}{\code{string} - Name of Table A on server}

\item{TableNameB.S}{\code{string} - Name of Table B on server}

\item{ByStatement.S}{\code{string} - The insides of a \code{dplyr::join_by()}-Statement defining how to join tables}

\item{JoinType.S}{\code{string} - Name of \code{dplyr::join}-function used, one of:
\itemize{\item 'left_join' (Default)
\item 'right_join'
\item 'full_join'
\item 'inner_join'}}
}
\value{
A \code{data.frame} resulting from join operation
}
\description{
Join tables on server, making use of \code{dplyr} mutating join operations.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
