% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadRawData.R
\name{LoadRawData}
\alias{LoadRawData}
\title{LoadRawData}
\usage{
LoadRawData(OpalDBConnection)
}
\value{
A list of tibbles containing raw data as described by the Raw Data Model
}
\description{
LoadRawData
}
\author{
Bastian Reiter
}
