% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTestDB.R
\name{MakeTestDB}
\alias{MakeTestDB}
\title{MakeTestDB}
\usage{
MakeTestDB(TestData)
}
\arguments{
\item{TestData}{A list of data frames representing tables to be loaded into the data base. Must have the structure of the current CCP Raw Data Model.}
}
\value{
A data bases connection object
}
\description{
Takes CCP test data, loads it into a virtual data base and returns the data base connection object.
}
\author{
Bastian Reiter
}
