% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapSystemicTherapyRegimen.R
\name{MapSystemicTherapyRegimen}
\alias{MapSystemicTherapyRegimen}
\title{MapSystemicTherapyRegimen}
\usage{
MapSystemicTherapyRegimen(
  InputData,
  Map.Regimens,
  Substances.MinimumCongruence = 0.8,
  Substances.AllowedCountDifferenceRange = c(0, 2),
  Substances.AllowAdditionals = FALSE
)
}
\arguments{
\item{InputData}{\code{data.frame} - Containing data on the following features:
\itemize{ \item  }}

\item{Map.Regimens}{\code{data.frame}}

\item{Substances.MinimumCongruence}{\code{double}}

\item{Substances.AllowedCountDifferenceRange}{\code{numeric vector}}

\item{Substances.AllowAdditionals}{\code{logical scalar}}
}
\value{
A \code{list}:
\itemize{ \item OutputData - The input \code{data.frame} with additional data
\item Tracker - A deidentified version of 'OutputData' for reporting }
}
\description{
Performs mapping from substances to regimen in systemic cancer therapy
}
\author{
Bastian Reiter
}
