% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapUICCStage.R
\name{MapUICCStage}
\alias{MapUICCStage}
\title{MapUICCStage}
\usage{
MapUICCStage(
  InputData,
  Map.TNMGroup,
  Map.UICCStage,
  AcceptableTNMCongruence = 0.8
)
}
\arguments{
\item{InputData}{\code{data.frame} - Containing data on the following features:
\itemize{ \item ICD10Code
\item ICD10Code.Short
\item ICDOTopographyCode
\item ICDOTopographyCode.Short
\item ICDOMorphologyCode
\item ICDOMorphologyCode.Short
\item ICDOMorphologyHistologyCode
\item Grading
\item TNM.T
\item TNM.N
\item TNM.M
\item TNMVersion }}

\item{Map.TNMGroup}{\code{data.frame}}

\item{Map.UICCStage}{\code{data.frame}}

\item{AcceptableTNMCongruence}{\code{double} - The threshold defining how much congruence in fuzzy matching is needed to accept a match. Given as a ratio number (Count of matching features in relation to number of all features used for matching). Must be a value between 0 and 1.}
}
\value{
The input \code{data.frame} with additional data
}
\description{
Performs mapping from TNM data to UICC stage according to TNM classification
}
\author{
Bastian Reiter
}
