% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchDataDS.R
\name{MatchDataDS}
\alias{MatchDataDS}
\title{MatchDataDS}
\usage{
MatchDataDS(
  DataFrame.S,
  GroupingVariable.S,
  CustomGroupingVariable.S = NULL,
  MatchingVariables.S,
  MatchItArguments.S = NULL
)
}
\arguments{
\item{DataFrame.S}{\code{string} - Name of \code{data.frame} containing source data}

\item{GroupingVariable.S}{\code{string} - A character string containing the \code{formula} used in \code{MatchIt::matchit()}. Example: 'Treatment ~ CovA + CovB + CovC + ...'. Note: The 'treatment variable' that is stated in the formula must be a binary variable.}

\item{CustomGroupingVariable.S}{\code{character} -}

\item{MatchingVariables.S}{\code{character} -}

\item{MatchItArguments.S}{\code{list} - Elements correspond to arguments of \code{MatchIt::matchit()}}
}
\value{
A ...
}
\description{
Obtain matched subsets from a given source \code{data.frame}, usually \code{ADS$Patients} from the Augmented Data Set (ADS).
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
