% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta.ADS}
\alias{Meta.ADS}
\title{CCP meta data: Table names, feature names and eligible values in AugmentedDataSet (ADS)}
\format{
\subsection{\code{Meta.ADS}}{

\code{tibble}
\describe{
\item{TableName}{Table name}
\item{FeatureName}{Feature name}
\item{FeatureType}{Feature data type}
\item{Value}{Eligible value}
\item{Label}{Label for values}
\item{HasEligibleValueSet}{logical}
\item{FeatureIsFromCDS}{logical}
}
}
}
\usage{
Meta.ADS
}
\description{
A \code{tibble} containing meta data
}
\author{
Bastian Reiter
}
\keyword{datasets}
