% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta.Features}
\alias{Meta.Features}
\title{CCP meta data: Features}
\format{
\subsection{\code{Meta.Features}}{

\code{tibble}
\describe{
\item{FeatureID}{}
\item{TableName.Raw}{Name of table in Opal}
\item{TableName.Curated}{Name of table after loading into R session}
\item{FeaturePosition}{Position of Feature in Table}
\item{FeatureName.Raw}{Feature name in Raw Data Model}
\item{FeatureName.Curated}{Corresponding feature name in Curated Data Model}
\item{IsPrimaryKey}{Indicating whether feature serves as primary key for corresponding table}
\item{Type}{Data type}
\item{Scale}{Scale of measure}
\item{HasEligibleValueSet}{Indicating whether values of feature are part of a finite, discrete eligible value set}
\item{IsDiscriminatory}{Indicating whether feature is used to strictly discriminate between different entries. Used for discrimination of table entries in redundancy classification.}
\item{IsEssential}{Indicating whether feature holds essential information. Used for discrimination of table entries in redundancy classification.}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta.Features
}
\description{
A tibble containing meta data about corresponding feature names in Raw Data Model (RDM) and Curated Data Model (CDM)
}
\author{
Bastian Reiter
}
\keyword{datasets}
