% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta.Links}
\alias{Meta.Links}
\title{CCP meta data: Links}
\format{
\subsection{\code{Meta.Links}}{

\code{tibble}
\describe{
\item{LinkID}{}
\item{RelationshipID}{}
\item{ParentTableName}{Name of parent table}
\item{ChildTableName}{Name of child table}
\item{ParentFeatureID}{ID of parent feature}
\item{ParentFeatureName}{Name of parent feature}
\item{ChildFeatureID}{ID of child feature}
\item{ChildFeatureName}{Name of child feature}
\item{Check.ParentTableName}{For checking purposes during meta data entry}
\item{Check.ChildTableName}{For checking purposes during meta data entry}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta.Links
}
\description{
A tibble containing meta data about features used for linkage of related tables
}
\author{
Bastian Reiter
}
\keyword{datasets}
