% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta.Relationships}
\alias{Meta.Relationships}
\title{CCP meta data: Relationships}
\format{
\subsection{\code{Meta.Relationships}}{

\code{tibble}
\describe{
\item{RelationshipID}{}
\item{Type}{Type of relationship between tables}
\item{ParentTableID}{The ID of the 'parent' table in a relationship}
\item{ParentTableName}{The corresponding name of the 'parent' table in a relationship}
\item{ChildTableID}{The ID of the 'child' table in a relationship}
\item{ChildTableName}{The corresponding name of the 'child' table in a relationship}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta.Relationships
}
\description{
A tibble containing meta data about relationships between tables
}
\author{
Bastian Reiter
}
\keyword{datasets}
