% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_DataHarmonization}
\alias{Meta_DataHarmonization}
\title{Meta data on rule-based transformation of raw data}
\format{
\subsection{\code{Meta_DataHarmonization}}{

Tibble
\describe{
\item{Profile}{}
\item{Table}{}
\item{Feature}{}
\item{EvaluationOrder}{}
\item{Operation}{}
\item{LookupValue}{}
\item{NewValue}{}
\item{Comment}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_DataHarmonization
}
\description{
A tibble containing rules
}
\author{
Bastian Reiter
}
\keyword{datasets}
