% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_DiagnosisRedundancy}
\alias{Meta_DiagnosisRedundancy}
\title{Meta data on rule-based identification of redundant diagnosis entries}
\format{
\subsection{\code{Meta_DiagnosisRedundancy}}{

Tibble
\describe{
\item{Profile}{}
\item{Feature}{}
\item{Value}{}
\item{ValueRank}{}
\item{EvaluationOrder}{}
\item{Condition}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_DiagnosisRedundancy
}
\description{
A tibble containing rules
}
\author{
Bastian Reiter
}
\keyword{datasets}
