% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_FeatureNames}
\alias{Meta_FeatureNames}
\title{CCP Meta Data: Feature names}
\format{
\subsection{\code{Meta_FeatureNames}}{

Tibble
\describe{
\item{TableName_Harmonized}{Table selector}
\item{FeaturePosition}{Position of Feature in Table}
\item{FeatureName_Raw}{Feature name in Raw Data Model}
\item{FeatureName_Harmonized}{Feature name in Harmonized Data Model}
}
}
}
\source{
\url{https://github.com/BastianReiter/CCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_FeatureNames
}
\description{
A tibble containing meta data about feature names in Raw Data Model (RDM) and Harmonized Data Model (HDM)
}
\keyword{datasets}
