% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_FeatureObligations}
\alias{Meta_FeatureObligations}
\title{Containing preferences on which features in CDS should be considered obligatory}
\format{
\subsection{\code{Meta_FeatureObligations}}{

\code{tibble}
\describe{
\item{TableName}{}
\item{FeatureName}{}
\item{Default}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_FeatureObligations
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
