% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_Features}
\alias{Meta_Features}
\title{CCP meta data: Features}
\format{
\subsection{\code{Meta_Features}}{

Tibble
\describe{
\item{TableName_Raw}{Name of table in Opal}
\item{TableName_Curated}{Name of table after loading into R session}
\item{FeaturePosition}{Position of Feature in Table}
\item{FeatureName_Raw}{Feature name in Raw Data Model}
\item{FeatureName_Curated}{Corresponding feature name in Curated Data Model}
\item{IsPrimaryKey}{Indicating whether feature serves as primary key for corresponding table}
\item{IsObligatory}{Indicating whether feature is obligatory for eligibility}
\item{Type}{Data type}
\item{Scale}{Scale of measure}
\item{HasEligibleValueSet}{Indicating whether values of feature are part of a finite, discrete eligible value set}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_Features
}
\description{
A tibble containing meta data about corresponding feature names in Raw Data Model (RDM) and Curated Data Model (CDM)
}
\author{
Bastian Reiter
}
\keyword{datasets}
