% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_FeatureTracking}
\alias{Meta_FeatureTracking}
\title{Containing preferences on which features should be tracked/monitored during curation process}
\format{
\subsection{\code{Meta_FeatureTracking}}{

\code{tibble}
\describe{
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{Indicating whether values of feature are part of a finite, discrete eligible value set}
\item{Default}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_FeatureTracking
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
