% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_Tables}
\alias{Meta_Tables}
\title{CCP meta data: Tables}
\format{
\subsection{\code{Meta_Tables}}{

\code{tibble}
\describe{
\item{TableName_Raw}{Table name in Raw Data Model}
\item{TableName_Curated}{Corresponding table name in Curated Data Model}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_Tables
}
\description{
A tibble containing meta data about corresponding table names in Raw Data Model (RDM) and Curated Data Model (CDM)
}
\author{
Bastian Reiter
}
\keyword{datasets}
