% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_Values}
\alias{Meta_Values}
\title{CCP meta data: Values}
\format{
\subsection{\code{Meta_Values}}{

\code{tibble}
\describe{
\item{FeatureID}{}
\item{Table}{Table name}
\item{Feature}{Feature name}
\item{ScaleLevel}{Scale level of feature}
\item{Value_Raw}{Value in original / raw data}
\item{Value_Curated}{Value as preferred}
\item{Label_Curated}{Label for coded feature values}
\item{Label_Raw}{Label in original data}
\item{FactorRank}{Used to determine order in values}
\item{ComparatorCode}{Assignment of numeric value to certain non-numeric values to enable comparison operations}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_Values
}
\description{
A tibble containing meta data about data values
}
\author{
Bastian Reiter
}
\keyword{datasets}
