% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_ValueSets}
\alias{Meta_ValueSets}
\title{CCP Meta Data: Value sets}
\format{
\subsection{\code{Meta_ValueSets}}{

Tibble
\describe{
\item{TableName_Curated}{Table selector}
\item{FeaturePosition}{Position of Feature in Table}
\item{FeatureName}{Feature name}
\item{ScaleLevel}{Scale level of feature}
\item{Value_Raw}{Value in original / raw data}
\item{Value_Curated}{Value as preferred}
\item{Label_Curated}{Label for coded feature values}
\item{Label_Raw}{Label in original data}
\item{FactorRank}{Used to determine order in values}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_ValueSets
}
\description{
A tibble containing meta data about eligible data values in Curated Data Model (CDM)
}
\keyword{datasets}
