% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MutateTableDS.R
\name{MutateTableDS}
\alias{MutateTableDS}
\title{MutateTableDS}
\usage{
MutateTableDS(TableName.S, MutateExpression.S, GroupBy.S = NULL)
}
\arguments{
\item{TableName.S}{\code{string} - Name of \code{data.frame}}

\item{MutateExpression.S}{\code{string} - \code{dplyr::mutate} expression as string}

\item{GroupBy.S}{\code{string} - Optional \code{dplyr::group_by} expression as string}
}
\value{
A \code{data.frame} resulting from mutating operation
}
\description{
Make use of \code{dplyr::mutate()} to create new features in a given table (\code{data.frame}).
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
