% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizeTable.R
\name{NormalizeTable}
\alias{NormalizeTable}
\title{NormalizeTable}
\usage{
NormalizeTable(DataFrame, TableName, RuleSet, RuleSet.Profile)
}
\arguments{
\item{DataFrame}{\code{data.frame} containing data to be transformed}

\item{TableName}{\code{string} - Name of the table (to enable mapping to normalization rules)}

\item{RuleSet}{\code{data.frame} - Contains predefined set of normalization rules}

\item{RuleSet.Profile}{\code{string} - Profile name stated in 'RuleSet'}
}
\value{
The transformed input \code{data.frame}
}
\description{
Carries out table harmonization procedures (like 'splitting/expanding') based on a given rule set.
}
\details{
Uses \code{tidyr} functionality to transform table based on normalization rules defined in a given rule set (Default: dsCCPhos::Meta_TableHarmonization)
}
\author{
Bastian Reiter
}
