% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Proc.EventFeatures}
\alias{Proc.EventFeatures}
\title{Proc.EventFeatures}
\format{
\subsection{\code{Proc.EventFeatures}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{Feature}{}
\item{Value}{}
\item{ValueRank}{}
\item{EvaluationOrder}{}
\item{Condition}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Proc.EventFeatures
}
\description{
Processing rules for engineering of informative features in event data
}
\author{
Bastian Reiter
}
\keyword{datasets}
