% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Proc.TableNormalization}
\alias{Proc.TableNormalization}
\title{Processing preferences on table normalization operations}
\format{
\subsection{\code{Proc.TableNormalization}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{EvaluationOrder}{}
\item{Operation}{}
\item{Comment}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Proc.TableNormalization
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
