% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecodeData.R
\name{RecodeData}
\alias{RecodeData}
\title{RecodeData}
\usage{
RecodeData(TargetVector, Dictionary)
}
\arguments{
\item{TargetVector}{Vector that recoding is performed on}

\item{Dictionary}{A named character vector. Vector names are look-up values, vector values are substitute values.}
}
\value{
A vector
}
\description{
Uses a dictionary in the form of a named vector to perform recoding on a target vector.
}
\author{
Bastian Reiter
}
