% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceDiagnosisIDs.R
\name{ReplaceDiagnosisIDs}
\alias{ReplaceDiagnosisIDs}
\title{ReplaceDiagnosisIDs}
\usage{
ReplaceDiagnosisIDs(Table, IDMapping)
}
\arguments{
\item{Table}{Table from partially curated CCP data set}

\item{IDMapping}{Data frame with mapping of DiagnosisIDs to be replaced}
}
\value{
Data frame with replaced DiagnosisIDs
}
\description{
After classification of redundant and associated diagnosis entries, replace DiagnosisIDs in related table.
}
\author{
Bastian Reiter
}
