% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RuleSet_DiagnosisAssociation}
\alias{RuleSet_DiagnosisAssociation}
\title{Meta data on rule-based identification of associations between diagnosis entries}
\format{
\subsection{\code{RuleSet_DiagnosisAssociation}}{

Tibble
\describe{
\item{Profile}{}
\item{Feature}{}
\item{Value}{}
\item{ValueRank}{}
\item{EvaluationOrder}{}
\item{BlockConnector}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
RuleSet_DiagnosisAssociation
}
\description{
A tibble containing rules
}
\author{
Bastian Reiter
}
\keyword{datasets}
