% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RuleSet_RawDataTransformation}
\alias{RuleSet_RawDataTransformation}
\title{Meta data on rule-based transformation of raw data}
\format{
\subsection{\code{RuleSet_RawDataTransformation}}{

Tibble
\describe{
\item{Profile}{}
\item{Table}{}
\item{Feature}{}
\item{EvaluationOrder}{}
\item{Operation}{}
\item{REP_OriginalValue}{}
\item{REP_NewValue}{}
\item{Comment}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
RuleSet_RawDataTransformation
}
\description{
A tibble containing rules
}
\author{
Bastian Reiter
}
\keyword{datasets}
