% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.CurationProcess}
\alias{Set.CurationProcess}
\title{Settings on how data set curation is supposed to proceed}
\format{
\subsection{\code{Set.CurationProcess}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{TableID}{}
\item{Table}{}
\item{PrimaryTableCleaning}{}
\item{TableNormalization}{}
\item{DataRemediation}{}
\item{DataRemediation.NonconformingRecords}{}
\item{SecondaryTableCleaning}{}
\item{RecordSubsumption}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.CurationProcess
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
