% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.DataHarmonization}
\alias{Set.DataHarmonization}
\title{Settings on which feature should be harmonized and which methods to use for each feature}
\format{
\subsection{\code{Set.DataHarmonization}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{}
\item{RunHarmonization}{}
\item{HarmonizationOrder}{}
\item{Method.TransformativeExpressions}{}
\item{Method.Dictionary}{}
\item{Method.FuzzyStringMatching}{}
\item{Method.NaiveBayes}{}
\item{MatchToDictionaryLookupsInFSM}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.DataHarmonization
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
