% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.DiagnosisAssociation}
\alias{Set.DiagnosisAssociation}
\title{Settings on rule-based identification of associations between diagnosis entries}
\format{
\subsection{\code{Set.DiagnosisAssociation}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{Feature}{}
\item{Value}{}
\item{ValueRank}{}
\item{EvaluationOrder}{}
\item{Condition}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.DiagnosisAssociation
}
\description{
A tibble containing rules
}
\author{
Bastian Reiter
}
\keyword{datasets}
