% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.DiagnosisRedundancy}
\alias{Set.DiagnosisRedundancy}
\title{Settings on rule-based identification of redundant diagnosis entries}
\format{
\subsection{\code{Set.DiagnosisRedundancy}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{Feature}{}
\item{Value}{}
\item{ValueRank}{}
\item{EvaluationOrder}{}
\item{Condition}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.DiagnosisRedundancy
}
\description{
A tibble containing rules
}
\author{
Bastian Reiter
}
\keyword{datasets}
