% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.Dictionary}
\alias{Set.Dictionary}
\title{Dictionary data used in Data Harmonization}
\format{
\subsection{\code{Set.Dictionary}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{LookupValue}{}
\item{NewValue}{}
\item{Comment}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.Dictionary
}
\description{
A tibble containing look-up values and corresponding replacements
}
\author{
Bastian Reiter
}
\keyword{datasets}
