% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.FeatureObligations}
\alias{Set.FeatureObligations}
\title{Containing preferences on which features in CDS should be considered obligatory}
\format{
\subsection{\code{Set.FeatureObligations}}{

\code{tibble}
\describe{
\item{Table}{}
\item{Feature}{}
\item{Default}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.FeatureObligations
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
