% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.FeatureRequirements}
\alias{Set.FeatureRequirements}
\title{Containing requirements for table features in CDS}
\format{
\subsection{\code{Set.FeatureRequirements}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{Type}{}
\item{Scale}{}
\item{Availability}{}
\item{NegligibleValues}{}
\item{ValidityCriteria}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.FeatureRequirements
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
