% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.FeatureTracking}
\alias{Set.FeatureTracking}
\title{Settings on which features should be tracked/monitored during curation process}
\format{
\subsection{\code{Set.FeatureTracking}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{Indicating whether values of feature are part of a finite, discrete eligible value set}
\item{IsTracked}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.FeatureTracking
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
