% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.FuzzyStringMatching}
\alias{Set.FuzzyStringMatching}
\title{Set.FuzzyStringMatching}
\format{
\subsection{\code{Set.FuzzyStringMatching}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{}
\item{PreferredMethod}{}
\item{FindBestMethod}{}
\item{Tolerance}{}
\item{Preprocessing.FlattenCase}{}
\item{Preprocessing.RemoveAllWhiteSpace}{}
\item{Preprocessing.SquishWhiteSpace}{}
\item{Stringdist.useBytes}{}
\item{Stringdist.weight.d}{}
\item{Stringdist.weight.i}{}
\item{Stringdist.weight.s}{}
\item{Stringdist.weight.t}{}
\item{Stringdist.q}{}
\item{Stringdist.p}{}
\item{Stringdist.bt}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.FuzzyStringMatching
}
\description{
Feature-specific settings for Fuzzy String Matching
}
\author{
Bastian Reiter
}
\keyword{datasets}
