% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.PrimaryTableCleaning}
\alias{Set.PrimaryTableCleaning}
\title{Settings on primary table cleaning procedures}
\format{
\subsection{\code{Set.PrimaryTableCleaning}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{TableID}{}
\item{Table}{}
\item{UnlinkedRecords.Detect}{}
\item{UnlinkedRecords.Remove}{}
\item{EmptyStrings.Detect}{}
\item{EmptyStrings.Substitute}{}
\item{EmptyStrings.Substitution}{}
\item{DuplicateRecords.Detect}{}
\item{DuplicateRecords.Remove}{}
\item{FeatureAvailabilityViolations.Detect}{}
\item{FeatureAvailabilityViolations.Remove}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.PrimaryTableCleaning
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
