% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.RecordSubsumption}
\alias{Set.RecordSubsumption}
\title{Settings on record subsumption}
\format{
\subsection{\code{Set.RecordSubsumption}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{TableID}{}
\item{Table}{}
\item{SubsumptionRedundancies.Detect}{}
\item{SubsumptionRedundancies.Remove}{\code{logical flag} - Indicates whether to remove table records that are considered redundant by subsumption. If set to \code{FALSE}, the redundant records are marked as such and preserved.}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.RecordSubsumption
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
