% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.SecondaryTableCleaning}
\alias{Set.SecondaryTableCleaning}
\title{Settings on secondary table cleaning procedures}
\format{
\subsection{\code{Set.SecondaryTableCleaning}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{TableID}{}
\item{Table}{}
\item{EmptyStrings.Detect}{}
\item{EmptyStrings.Substitute}{}
\item{EmptyStrings.Substitution}{}
\item{DuplicateRecords.Detect}{}
\item{DuplicateRecords.Remove}{}
\item{FeatureAvailabilityViolations.Detect}{}
\item{FeatureAvailabilityViolations.Remove}{}
\item{SecondaryRedundancies.Detect}{}
\item{SecondaryRedundancies.Remove}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.SecondaryTableCleaning
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
