% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.StringExtraction}
\alias{Set.StringExtraction}
\title{Set.StringExtraction}
\format{
\subsection{\code{Set.StringExtraction}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{}
\item{WorkStringMinimumLength}{}
\item{RefStringMinimumLength}{}
\item{StringLengthMinimumRatio}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.StringExtraction
}
\description{
Feature-specific settings for String Extraction used in Data Remediation
}
\author{
Bastian Reiter
}
\keyword{datasets}
