% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.TransformativeExpressions}
\alias{Set.TransformativeExpressions}
\title{Settings on transformative expressions used in Data Harmonization}
\format{
\subsection{\code{Set.TransformativeExpressions}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{EvaluationOrder}{}
\item{Expression}{}
\item{Comment}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.TransformativeExpressions
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
