% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeEventData.R
\name{SummarizeEventData}
\alias{SummarizeEventData}
\title{SummarizeEventData}
\usage{
SummarizeEventData(EventEntries, ProgressBarObject = NULL)
}
\arguments{
\item{EventEntries}{Data frame that contains entries from unnested ADS_Events}

\item{ProgressBarObject}{Optionally pass object of type progress::progress_bar to display progress}
}
\value{
Data frame containing summarized information
}
\description{
Auxiliary function within \code{\link{AugmentDataDS}}
}
\details{
Summarize data stored in ADS_Events and create meaningful features for ADS_Diagnosis and ADS_Patient
}
\author{
Bastian Reiter
}
