% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrackValueCounts.R
\name{TrackValueCounts}
\alias{TrackValueCounts}
\title{TrackValueCounts}
\usage{
TrackValueCounts(DataFrame, Features, TransformationStage = NULL)
}
\arguments{
\item{DataFrame}{A data frame}

\item{Features}{A list of vectors. Name of vector object gives name of feature, optional vector values give eligible values.}

\item{TransformationStage}{String representing transformation stage (usually "Raw" / "After Transformation" / "After Exclusion")}
}
\value{
A tibble
}
\description{
Track occurrence of a feature's unique values, their eligibility and frequency
}
\author{
Bastian Reiter
}
