% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{TransformData}
\usage{
TransformData(DataFrame, TableName, RuleSet, RuleProfile)
}
\arguments{
\item{TargetVector}{Vector that recoding is performed on}

\item{Dictionary}{A named character vector. Vector names are look-up values, vector values are substitute values.}
}
\value{
A vector
}
\description{
Uses dplyr::mutate() to transform data based on transformation rules defined in dsCCPhos::RuleSet_RawDataTransformation
}
\author{
Bastian Reiter
}
