% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnpackAugmentedDataSetDS.R
\name{UnpackAugmentedDataSetDS}
\alias{UnpackAugmentedDataSetDS}
\title{UnpackAugmentedDataSetDS}
\usage{
UnpackAugmentedDataSetDS(
  Name_AugmentationOutput = "AugmentationOutput",
  TableName
)
}
\arguments{
\item{Name_AugmentationOutput}{String | Name of the list object created by AugmentDataDS() | Default: 'AugmentationOutput'}

\item{TableName}{String | Name of Augmented Data Set table in AugmentationOutput}
}
\value{
A data frame (table from Augmented Data Set)
}
\description{
Make table within Augmented Data Set directly addressable by assigning it to a data frame object in R server session
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
