% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnpackCuratedDataSetDS.R
\name{UnpackCuratedDataSetDS}
\alias{UnpackCuratedDataSetDS}
\title{UnpackCuratedDataSetDS}
\usage{
UnpackCuratedDataSetDS(Name_CurationOutput = "CurationOutput", TableName)
}
\arguments{
\item{Name_CurationOutput}{String | Name of the list object created by CurateDataDS() | Default: 'CurationOutput'}

\item{TableName}{String | Name of Curated Data Set table in CurationOutput as defined in \code{\link{Meta_TableNames}}}
}
\value{
A data frame (table from Curated Data Set)
}
\description{
Make table within Curated Data Set directly addressable by assigning it to a data frame object in R server session
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
