% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XMLtoDB.R
\name{XMLtoDB}
\alias{XMLtoDB}
\title{XMLtoDB}
\usage{
XMLtoDB(
  XMLSchemaFilePath = "./Development/MetaData/SchemaOpalDB.xml",
  TableNodeElement = "container",
  TableNameAttribute = "opal-table",
  ColumnNodeElement = "attribute",
  ColumnNameAttribute = "csv-column",
  ColumnTypeAttribute = "opal-value-type",
  DBPath = ":memory:"
)
}
\arguments{
\item{DBPath}{}
}
\value{
A data base connection object
}
\description{
Make (virtual) data base from XML Schema
}
