% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentDataDS.R
\name{AugmentDataDS}
\alias{AugmentDataDS}
\title{AugmentDataDS}
\usage{
AugmentDataDS(Name_CurationOutput = "CurationOutput")
}
\arguments{
\item{Name_CurationOutput}{String | Name of the list object created by CurateDataDS() | Default: 'CurationOutput'}
}
\description{
Transforms Curated Data Set (CDS) into Augmented Data Set (ADS)
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
