% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AugmentDataDS.R
\name{AugmentDataDS}
\alias{AugmentDataDS}
\title{AugmentDataDS}
\usage{
AugmentDataDS(
  CuratedDataSetName.S = "CuratedDataSet",
  Settings.S = list(CutoffValues = list(DaysDiagnosisToInitialStaging = 30),
    DiagnosisAssociation = list(Check = TRUE, RuleSet =
    dsCCPhos::Meta_DiagnosisAssociation, Profile = "Default"), EventFeatures =
    list(RuleSet = dsCCPhos::Meta_EventFeatures, Profile = "Default"), OverallSurvival =
    list(ReferenceEvent = c(EventClass = "Diagnosis", EventSubclass =
    "InitialDiagnosis")), TherapyOfInterest = list(EventSubclass = "Surgery",
    EventSubclassRank = 1), TimeToEvent = list(ReferenceEvent = list(EventClass =
    "Diagnosis", 
     EventSubclass = "InitialDiagnosis"), TargetEvent = list(EventClass
    = "VitalStatus", EventSubclass = "Deceased")))
)
}
\arguments{
\item{CuratedDataSetName.S}{\code{character} - Name of the Curated Data Set object on server - Default: 'CuratedDataSet'}

\item{Settings.S}{\code{list} - Settings passed to function
\itemize{\item DiagnosisAssociation \code{list}
\itemize{\item Check \code{logical} - Whether or not to classify associated diagnosis entries
\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_DiagnosisAssociation}
\item Profile \code{character} - Profile name defining rule set to be used for classification of diagnosis associations. Profile name must be stated in \code{DiagnosisAssociation$RuleSet} - Default: 'Default'}
\item EventFeatures \code{list}
\itemize{\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_EventFeatures}
\item Profile \code{character} - Profile name defining rule set to be used for event feature engineering. Profile name must be stated in \code{EventFeatures$RuleSet} - Default: 'Default'}}}
}
\value{
A \code{list} containing the following objects:
\itemize{\item AugmentedDataSet \code{list}
\itemize{\item Events
\item DiseaseCourse
\item Therapy
\item Diagnosis
\item Patient}
\item AugmentationReport \code{list}
\item AugmentationMessages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Transforms Curated Data Set (CDS) into Augmented Data Set (ADS)
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
