% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.AugmentDataDS.R
\name{CCP.AugmentDataDS}
\alias{CCP.AugmentDataDS}
\title{CCP.AugmentDataDS}
\usage{
CCP.AugmentDataDS(
  CuratedDataSetName.S = "CCP.CuratedDataSet",
  CutoffValues.DaysDiagnosisToInitialStaging.S = 30,
  DiagnosisAssociation.Check.S = TRUE,
  DiagnosisAssociation.RuleSet.S = dsCCPhos::Set.DiagnosisAssociation,
  DiagnosisAssociation.Profile.S = "Default",
  EventFeatures.RuleSet.S = dsCCPhos::Proc.EventFeatures,
  EventFeatures.Profile.S = "Default",
  Imputation.SystemicTherapyRegimen.Run.S = TRUE,
  Imputation.SystemicTherapyRegimen.AcceptableSubstanceCongruence.S = 1,
  Imputation.SystemicTherapyRegimen.RunAccuracyTest.S = TRUE,
  Imputation.UICCStage.Run.S = TRUE,
  Imputation.UICCStage.AcceptableTNMCongruence.S = 0.8,
  Imputation.UICCStage.RunAccuracyTest.S = TRUE,
  OverallSurvival.ReferenceEvent.EventClass.S = "Diagnosis",
  OverallSurvival.ReferenceEvent.EventSubclass.S = "InitialDiagnosis",
  TherapyOfInterest.EventSubclass.S = "Surgery",
  TherapyOfInterest.EventSubclassRank.S = 1,
  TimeToEvent.ReferenceEvent.EventClass.S = "Diagnosis",
  TimeToEvent.ReferenceEvent.EventSubclass.S = "InitialDiagnosis",
  TimeToEvent.TargetEvent.EventClass.S = "VitalStatus",
  TimeToEvent.TargetEvent.EventSubclass.S = "Deceased",
  MetaDataKey.S = Sys.getenv("METADATAKEY")
)
}
\arguments{
\item{CuratedDataSetName.S}{\code{character} - Name of the Curated Data Set object on server - Default: 'CuratedDataSet'}

\item{Settings.S}{\code{list} - Settings passed to function
\itemize{\item DiagnosisAssociation \code{list}
\itemize{\item Check \code{logical} - Whether or not to classify associated diagnosis entries
\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Set.DiagnosisAssociation}
\item Profile \code{character} - Profile name defining rule set to be used for classification of diagnosis associations. Profile name must be stated in \code{DiagnosisAssociation$RuleSet} - Default: 'Default'}
\item EventFeatures \code{list}
\itemize{\item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Proc.EventFeatures}
\item Profile \code{character} - Profile name defining rule set to be used for event feature engineering. Profile name must be stated in \code{EventFeatures$RuleSet} - Default: 'Default'}}}
}
\value{
A \code{list} containing the following objects:
\itemize{\item AugmentedDataSet \code{list}
\itemize{\item Events
\item DiseaseCourse
\item Therapy
\item Diagnosis
\item Patient}
\item AugmentationReport \code{list}
\item Messages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} \cr\cr
Transforms Curated Data Set (CDS) into Augmented Data Set (ADS)
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
