% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCP.GetCohortDescriptionDS.R
\name{CCP.GetCohortDescriptionDS}
\alias{CCP.GetCohortDescriptionDS}
\title{CCP.GetCohortDescriptionDS}
\usage{
CCP.GetCohortDescriptionDS(
  DataSetName.S = "AugmentedDataSet",
  Stage.S = "Augmented"
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{Stage.S}{\code{string} - Indicating the transformation stage of the described data set, one of "Raw" / "Curated" / "Augmented" - Default: "Augmented"}
}
\value{
\code{list}
}
\description{
What it does
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
