% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisAssociations.R
\name{ClassifyDiagnosisAssociations}
\alias{ClassifyDiagnosisAssociations}
\title{ClassifyDiagnosisAssociations}
\usage{
ClassifyDiagnosisAssociations(
  DiagnosisEntries,
  Method = "Rule-based",
  RulesProfile = "Default"
)
}
\arguments{
\item{DiagnosisEntries}{Data frame (partially curated) that contains multiple diagnosis entries}

\item{Method}{String | Determines the method being used for classification
\itemize{
\item "Rule-based" (Default)
\item "ML-Model" }}

\item{ClassificationRules}{}
}
\value{
Data frame enhanced with classified diagnosis associations
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
In patients with multiple diagnosis entries: Identify and classify associations between diagnosis entries.
}
\author{
Bastian Reiter
}
