% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassifyDiagnosisRedundancy.R
\name{ClassifyDiagnosisRedundancy}
\alias{ClassifyDiagnosisRedundancy}
\title{ClassifyDiagnosisRedundancy}
\usage{
ClassifyDiagnosisRedundancy(
  DiagnosisEntries,
  RuleCalls,
  ProgressBarObject = NULL
)
}
\arguments{
\item{DiagnosisEntries}{\code{data.frame} that (usually) contains multiple diagnosis entries}

\item{RuleCalls}{\code{list} of unevaluated \code{dplyr::case_when}-Statements compiled from rules defined in \code{RuleSet_DiagnosisRedundancy}}

\item{ProgressBarObject}{\code{progress::progress_bar} object - Optionally pass progress bar object to display progress}
}
\value{
\code{data.frame} of consolidated diagnosis entries
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \cr\cr
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
In patients with multiple diagnosis entries: Identify, classify and remove redundant diagnosis entries.
}
\author{
Bastian Reiter
}
