% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanTable.R
\name{CleanTable}
\alias{CleanTable}
\title{CleanTable}
\usage{
CleanTable(
  Table,
  TableNameLookup,
  RemoveRedundantEntries = TRUE,
  FeatureObligations_RuleSet,
  FeatureObligations_Profile
)
}
\arguments{
\item{Table}{\code{data.frame} or \code{tibble} - The table object to be cleaned}

\item{TableNameLookup}{\code{character} - The table name(s) to be looked up in 'FeatureObligations_RuleSet'. Can be a single string or a character vector.}

\item{RemoveRedundantEntries}{\code{logical} - Whether redundant entries should be removed}

\item{FeatureObligations_RuleSet}{\code{data.frame}}

\item{FeatureObligations_Profile}{\code{character}}
}
\value{
\code{tibble} - Clean table
}
\description{
Auxiliary function in \code{dsCCPhos::CurateDataDS()} to perform exclusion of invalid table entries.
}
\author{
Bastian Reiter
}
