% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileClassificationCall.R
\name{CompileClassificationCall}
\alias{CompileClassificationCall}
\title{CompileClassificationCall}
\usage{
CompileClassificationCall(
  TargetFeature,
  RuleSet,
  RuleProfile,
  ValueIfNoRuleMet = NA
)
}
\arguments{
\item{TargetFeature}{\code{string} - Name of feature that is being engineered}

\item{RuleSet}{\code{data.frame} - Contains predefined set of rules}

\item{RuleProfile}{\code{string} - Profile name stated in rule set data frame}

\item{ValueIfNoRuleMet}{\code{string} - Value (as string) of TargetFeature if no rules are fulfilled | Default = "NA"}
}
\value{
\code{string} containing an unevaluated dplyr::case_when() expression
}
\description{
Auxiliary function to compile calls used in \code{\link{ClassifyDiagnosisAssociation}}
}
\details{
Based on a predefined rule set data frame, compile a dplyr::case_when() expression and return it as a string
}
\author{
Bastian Reiter
}
