% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileHarmonizationRules.R
\name{CompileHarmonizationRules}
\alias{CompileHarmonizationRules}
\title{CompileHarmonizationRules}
\usage{
CompileHarmonizationRules(TableName, RuleSet, RuleProfile)
}
\arguments{
\item{TableName}{String | Name of table that contains features to be transformed}

\item{RuleSet}{Data frame | Contains predefined set of harmonization rules}

\item{RuleProfile}{String | Profile name stated in rule set data frame}
}
\value{
String containing arguments for dplyr::mutate()
}
\description{
Auxiliary function within \code{\link{CurateDataDS}} or \code{\link{HarmonizeData}} respectively
}
\details{
Based on a predefined rule set data frame, compile argument expressions for dplyr::mutate()
}
\author{
Bastian Reiter
}
