% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompileHashTables.R
\name{CompileHashTables}
\alias{CompileHashTables}
\title{CompileHashTables}
\usage{
CompileHashTables(TableName, RuleSet, RuleProfile)
}
\arguments{
\item{TableName}{String | Name of table that contains features to be transformed}

\item{RuleSet}{Data frame | Contains predefined set of transformation rules}

\item{RuleProfile}{String | Profile name stated in rule set data frame}
}
\value{
List of named vectors (serving as hash tables)
}
\description{
Auxiliary function within \code{\link{CurateDataDS}} or \code{\link{TransformData}} respectively
}
\details{
Based on a predefined rule set data frame, compile hash tables for direct value transformation
}
\author{
Bastian Reiter
}
