% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountDeviations.R
\name{CountDeviations}
\alias{CountDeviations}
\title{CountDeviations}
\usage{
CountDeviations(
  ReferenceEntry,
  CandidateEntry,
  ComparedFeatures = colnames(ReferenceEntry)
)
}
\arguments{
\item{ReferenceEntry}{A data frame row}

\item{CandidateEntry}{A data frame row}

\item{ComparedFeatures}{Vector}
}
\value{
Integer number
}
\description{
Take two rows from a data frame and count deviating values
}
\author{
Bastian Reiter
}
