% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(Name_RawData = "RawData")
}
\arguments{
\item{Name_RawData}{String | Name of raw data object (list) on server | Default: 'RawData'}
}
\value{
A list containing the curated data and a curation report
}
\description{
Takes raw data and curates it while tracing transformation operations.
}
\details{
Server-side ASSIGN method
}
