% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  Name_RawDataSet = "RawDataSet",
  RulesProfile_DiagnosisAssociation = "Default",
  RulesProfile_DiagnosisRedundancies = "Default"
)
}
\arguments{
\item{Name_RawDataSet}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}

\item{RulesProfile_DiagnosisAssociation}{String | Profile name defining rule set to be used for classification of diagnosis association | Default: 'Default'}

\item{RulesProfile_DiagnosisRedundancies}{String | Profile name defining rule set to be used for classification of diagnosis redundancies | Default: 'Default'}
}
\value{
A list containing the Curated Data Set (CDS) and a curation report.
}
\description{
Takes Raw Data Set (RDS) and transforms it into Curated Data Set (CDS) while tracing transformation operations.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
