% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  RawDataSetName.S = "RawDataSet",
  RuleProfile_RawDataTransformation.S = "Default",
  RuleProfile_DiagnosisRedundancy.S = "Default",
  RuleProfile_DiagnosisAssociation.S = "Default"
)
}
\arguments{
\item{RawDataSetName.S}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list containing the following objects:
\itemize{\item CuratedDataSet (list)
\item CurationReport (list)
\item CurationMessages (list)}
}
\description{
Takes Raw Data Set (RDS) and transforms it into Curated Data Set (CDS) while tracing data transformation.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
