% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  Name_RawDataSet = "RawDataSet",
  RuleProfile_DiagnosisAssociation = "Default",
  RuleProfile_DiagnosisRedundancy = "Default"
)
}
\arguments{
\item{Name_RawDataSet}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}
}
\value{
A list containing the Curated Data Set (CDS) and a curation report.
}
\description{
Takes Raw Data Set (RDS) and transforms it into Curated Data Set (CDS) while tracing transformation operations.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
