% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  RawDataSetName.S = "RawDataSet",
  RuleSet_RawDataHarmonization.S = dsCCPhos::RuleSet_RawDataHarmonization,
  RuleProfile_RawDataHarmonization.S = "Default",
  RuleSet_DiagnosisRedundancy.S = dsCCPhos::RuleSet_DiagnosisRedundancy,
  RuleProfile_DiagnosisRedundancy.S = "Default",
  RuleSet_DiagnosisAssociation.S = dsCCPhos::RuleSet_DiagnosisAssociation,
  RuleProfile_DiagnosisAssociation.S = "Default"
)
}
\arguments{
\item{RawDataSetName.S}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}

\item{RuleSet_RawDataHarmonization.S}{Data frame}

\item{RuleSet_DiagnosisRedundancy.S}{Data frame}

\item{RuleSet_DiagnosisAssociation.S}{Data frame}
}
\value{
A list containing the following objects:
\itemize{\item CuratedDataSet (list)
\itemize{\item BioSampling
\item Diagnosis
\item Histology
\item Metastasis
\item MolecularDiagnostics
\item Patient
\item Progress
\item RadiationTherapy
\item Staging
\item Surgery
\item SystemicTherapy}
\item CurationReport (list)
\itemize{\item IneligibleEntries (named vector)
\item Transformation (list of data frames)
\itemize{\item BioSampling
\item Diagnosis
\item ...}
\item DiagnosisClassification (named vector)}
\item CurationMessages (list)}
}
\description{
Takes Raw Data Set (RDS) and transforms it into Curated Data Set (CDS) while tracing data transformation.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
