% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  RawDataSetName.S = "RawDataSet",
  Settings.S = list(DataHarmonization_RuleSet = dsCCPhos::Meta_DataHarmonization,
    DataHarmonization_Profile = "Default", DiagnosisRedundancy_Check = TRUE,
    DiagnosisRedundancy_RuleSet = dsCCPhos::Meta_DiagnosisRedundancy,
    DiagnosisRedundancy_Profile = "Default", DiagnosisAssociation_Check = TRUE,
    DiagnosisAssociation_RuleSet = dsCCPhos::Meta_DiagnosisAssociation,
    DiagnosisAssociation_Profile = "Default", FeatureObligations_RuleSet =
    dsCCPhos::Meta_FeatureObligations, FeatureObligations_Profile = "Default")
)
}
\arguments{
\item{RawDataSetName.S}{\code{character} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{Settings.S}{\code{list} - Settings passed to function
\itemize{\item DataHarmonization_RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_DataHarmonization}
\item DataHarmonization_Profile \code{character} - Profile name defining rule set to be used for data harmonization. Profile name must be stated in \code{RawDataHarmonization_RuleSet} - Default: 'Default'
\item DiagnosisRedundancy_Check \code{logical} - Whether or not to check for redundant diagnosis entries
\item DiagnosisRedundancy_RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_DiagnosisRedundancy}
\item DiagnosisRedundancy_Profile \code{character} - Profile name defining rule set to be used for classification of diagnosis redundancies. Profile name must be stated in \code{DiagnosisRedundancy_RuleSet} - Default: 'Default'
\item DiagnosisAssociation_Check \code{logical} - Whether or not to classify associated diagnosis entries
\item DiagnosisAssociation_RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_DiagnosisAssociation}
\item DiagnosisAssociation_Profile \code{character} - Profile name defining rule set to be used for classification of diagnosis associations. Profile name must be stated in \code{DiagnosisAssociation_RuleSet} - Default: 'Default'
\item FeatureObligations_RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureObligations}
\item FeatureObligations_Profile \code{character} - Profile name defining strict and trans-feature rules for obligatory feature content. Profile name must be stated in \code{FeatureObligations_RuleSet} - Default: 'Default'}}
}
\value{
A \code{list} containing the following objects:
\itemize{\item CuratedDataSet \code{list}
\itemize{\item BioSampling
\item Diagnosis
\item GeneralCondition
\item Histology
\item Metastasis
\item MolecularDiagnostics
\item OtherClassification
\item Patient
\item Progress
\item RadiationTherapy
\item Staging
\item Surgery
\item SystemicTherapy
\item TherapyRecommendation}
\item CurationReport \code{list}
\itemize{\item EntryCounts \code{tibble}
\itemize{\item Table
\item InitialCount
\item ExcludedPrimary
\item AfterPrimaryExclusion
\item ExcludedSecondary
\item AfterSecondaryExclusion}
\item Transformation (list of lists)
\itemize{\item Monitors
\item EligibilityOverviews
\item ValueSetOverviews}
\item DiagnosisClassification
\itemize{\item DiagnosisRedundancies
\item PatientsWithDiagnosisRedundancies
\item DiagnosisAssociations
\item PatientsWithDiagnosisAssociations}}
\item CurationMessages \code{list}}
}
\description{
Takes Raw Data Set (RDS) and transforms it into Curated Data Set (CDS) while tracing data transformation.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
