% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CurateDataDS.R
\name{CurateDataDS}
\alias{CurateDataDS}
\title{CurateDataDS}
\usage{
CurateDataDS(
  RawDataSetName.S = "RawDataSet",
  Settings.S = list(DataHarmonization = list(Run = TRUE, Methods =
    dsCCPhos::Meta_DataHarmonizationMethods, TransformativeExpressions =
    dsCCPhos::Meta_TransformativeExpressions, TransformativeExpressions.Profile =
    "Default", Dictionary = dsCCPhos::Meta_Dictionary, Dictionary.Profile = "Default",
    FuzzyStringMatching = dsCCPhos::Meta_FuzzyStringMatching, FuzzyStringMatching.Profile
    = "Default"), FeatureObligations = list(RuleSet = dsCCPhos::Meta_FeatureObligations,
    RuleSet.Profile = "Default"), FeatureTracking = list(RuleSet =
    dsCCPhos::Meta_FeatureTracking, 
     RuleSet.Profile = "Default"), TableCleaning =
    list(Run = TRUE), TableNormalization = list(Run = TRUE, RuleSet =
    dsCCPhos::Meta_TableNormalization, RuleSet.Profile = "Default")),
  DataHarmonization.Run = NULL,
  DataHarmonization.Methods = NULL,
  DataHarmonization.TransformativeExpressions = NULL,
  DataHarmonization.TransformativeExpressions.Profile = NULL,
  DataHarmonization.Dictionary = NULL,
  DataHarmonization.Dictionary.Profile = NULL,
  DataHarmonization.FuzzyStringMatching = NULL,
  DataHarmonization.FuzzyStringMatching.Profile = NULL,
  FeatureObligations.RuleSet = NULL,
  FeatureObligations.RuleSet.Profile = NULL,
  FeatureTracking.RuleSet = NULL,
  FeatureTracking.RuleSet.Profile = NULL,
  TableCleaning.Run = NULL,
  TableNormalization.Run = NULL,
  TableNormalization.RuleSet = NULL,
  TableNormalization.RuleSet.Profile = NULL
)
}
\arguments{
\item{RawDataSetName.S}{\code{character} - Name of Raw Data Set object (list) on server - Default: 'RawDataSet'}

\item{Settings.S}{\code{list} - Settings passed to function
\itemize{  \item \emph{DataHarmonization} - \code{list}
\itemize{ \item Run \code{logical} - Whether or not to perform data harmonization - Default: \code{TRUE}
\item Methods \code{data.frame} - Default: \code{dsCCPhos::Meta_DataHarmonizationMethods}
\item TransformativeExpressions \code{data.frame} - Default: \code{dsCCPhos::Meta_TransformativeExpressions}
\item TransformativeExpressions.Profile \code{string} - Profile used in \emph{TransformativeExpressions} - Default: 'Default'
\item Dictionary \code{data.frame} - Default: \code{dsCCPhos::Meta_Dictionary}
\item Dictionary.Profile \code{string} - Profile used in \emph{Dictionary} - Default: 'Default'
\item FuzzyStringMatching \code{data.frame} - Default: \code{dsCCPhos::Meta_FuzzyStringMatching}
\item FuzzyStringMatching.Profile \code{string} - Profile used in \emph{FuzzyStringMatching} - Default: 'Default'}
\item \emph{FeatureObligations} - \code{list}
\itemize{ \item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureObligations}
\item RuleSet.Profile \code{string} - Profile name defining strict and trans-feature rules for obligatory feature content. Profile name must be stated in \code{FeatureObligations$RuleSet} - Default: 'Default'}
\item \emph{FeatureTracking} - \code{list}
\itemize{ \item RuleSet \code{data.frame} - Default: \code{dsCCPhos::Meta_FeatureTracking}
\item RuleSet.Profile \code{string} - Profile name defining which features should be tracked/monitored during curation process. Profile name must be stated in \code{FeatureTracking$RuleSet} - Default: 'Default'}
\item \emph{TableCleaning} - \code{list}
\itemize{ \item Run \code{logical} - Whether or not to perform table cleaning (removal of redundant and ineligible entries) - Default: \code{TRUE}}
\item \emph{TableNormalization} - \code{list}
\itemize{ \item Run \code{logical} - Whether or not to perform table normalization - Default: \code{TRUE}
\item RuleSet \code{data.frame} - Deault: \code{dsCCPhos::Meta_TableNormalization}
\item RuleSet.Profile \code{string} - Profile name defining rule set to be used for table normalization. Profile name must be stated in \code{TableNormalization$RuleSet} - Default: 'Default'}}}
}
\value{
A \code{list} containing the following objects:
\itemize{\item CuratedDataSet \code{list}
\itemize{ \item BioSampling
\item Diagnosis
\item DiseaseStatus
\item GeneralCondition
\item Histology
\item Metastasis
\item MolecularDiagnostics
\item OtherClassification
\item Patient
\item RadiationTherapy
\item Staging
\item Surgery
\item SystemicTherapy
\item TherapyRecommendation}
\item CurationReport \code{list}
\itemize{ \item EntryCounts \code{tibble}
\itemize{ \item Table
\item InitialCount
\item ExcludedPrimary
\item AfterPrimaryExclusion
\item ExcludedSecondary
\item AfterSecondaryExclusion
\item ExcludedSecondaryRedundancy
\item AfterSecondaryRedundancyExclusion}
\item Transformation (list of lists)
\itemize{ \item Monitors
\item EligibilityOverviews
\item ValueSetOverviews}}
\item CurationMessages \code{list}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}} \cr\cr
Transforms Raw Data Set (RDS) into Curated Data Set (CDS) while tracing data transformation.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
