% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcludeCuratedDataDS.R, R/ExcludeRawDataDS.R
\name{ExcludeCuratedDataDS}
\alias{ExcludeCuratedDataDS}
\title{ExcludeCuratedDataDS}
\usage{
ExcludeCuratedDataDS(Name_RawDataSet = "RawDataSet")

ExcludeCuratedDataDS(Name_RawDataSet = "RawDataSet")
}
\arguments{
\item{Name_RawDataSet}{String | Name of Raw Data Set object (list) on server | Default: 'RawDataSet'}

\item{Name_CurationOutput}{String | Name of the list object created by CurateDataDS() | Default: 'CurationOutput'}
}
\value{
The Curation Output with modified Curated Data Set

Modified Raw Data Set object (list)
}
\description{
Exclude data from Curated Data Set

Exclude data from Raw Data Set (RDS)
}
\details{
Server-side ASSIGN method

Server-side ASSIGN method
}
\author{
Bastian Reiter
}
