% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterTableDS.R
\name{FilterTableDS}
\alias{FilterTableDS}
\title{FilterTableDS}
\usage{
FilterTableDS(TableName.S, FilterStatement.S)
}
\arguments{
\item{TableName.S}{\code{string} - Name of Table A on server}

\item{FilterStatement.S}{\code{string} - The insides of a \code{dplyr::filter()}-Statement defining which rows to select/exclude}
}
\value{
A \code{data.frame} resulting from filter operation
}
\description{
Perform filtering on a server-side table, making use of \code{dplyr::filter()} operations.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
