% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterTableDS.R
\name{FilterTableDS}
\alias{FilterTableDS}
\title{FilterTableDS}
\usage{
FilterTableDS(TableName.S, FilterExpression.S, GroupBy.S = NULL)
}
\arguments{
\item{TableName.S}{\code{string} - Name of \code{data.frame}}

\item{FilterExpression.S}{\code{string} - \code{dplyr::filter} expression as string}

\item{GroupBy.S}{\code{string} - Optional \code{dplyr::group_by} expression as string}
}
\value{
A \code{data.frame} resulting from filter operation
}
\description{
Filter a table (\code{data.frame}) making use of \code{dplyr} and \code{stringr}.
}
\details{
Server-side ASSIGN method
}
\author{
Bastian Reiter
}
