% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalizeDataTransformation.R
\name{FinalizeDataTransformation}
\alias{FinalizeDataTransformation}
\title{FinalizeDataTransformation}
\usage{
FinalizeDataTransformation(
  TargetVector,
  EligibleValueSet,
  ExcludeIneligibleValues = TRUE,
  ConvertToFactor = FALSE,
  AssignFactorLabels = FALSE
)
}
\arguments{
\item{TargetVector}{\code{vector} - Feature to be finalized}

\item{EligibleValueSet}{\code{data.frame} - Containing set of eligible values}

\item{ExcludeIneligibleValues}{\code{logical} - Whether to set all ineligible values NA | Default: TRUE}

\item{ConvertToFactor}{\code{logical} - Whether to convert vector to factor | Default: FALSE}

\item{AssignFactorLabels}{\code{logical} - Whether to assign factor labels during factor conversion | Default: FALSE}
}
\value{
A \code{vector} (or factor)
}
\description{
Auxiliary function within \code{\link{CurateDataDS}}
}
\details{
Ineligible data (including data that could not be transformed) is turned into NA per default.
Optional factor conversion establishes level order where appropriate.
Value eligibility and factor information is defined in dsCCPhos::Meta_ValueSets.
}
\author{
Bastian Reiter
}
