% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortDescriptionDS.R
\name{GetCohortDescriptionDS}
\alias{GetCohortDescriptionDS}
\title{GetCohortDescriptionDS}
\usage{
GetCohortDescriptionDS(Name_AugmentationOutput = "AugmentationOutput")
}
\arguments{
\item{Name_AugmentationOutput}{String | Name of object on server previously assigned by dsCCPhos::AugmentData() | Default: 'AugmentationOutput'}
}
\description{
What it does
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
