% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCohortDescriptionDS.R
\name{GetCohortDescriptionDS}
\alias{GetCohortDescriptionDS}
\title{GetCohortDescriptionDS}
\usage{
GetCohortDescriptionDS(
  DataSetName.S = "AugmentedDataSet",
  CCPDataSetType.S = "ADS"
)
}
\arguments{
\item{DataSetName.S}{\code{string} - Name of Data Set object (list) on server, usually "RawDataSet", "CuratedDataSet" or "AugmentedDataSet"}

\item{CCPDataSetType.S}{\code{string} - Indicating the type of CCP data set that should be described, one of "RDS" / "CDS" / "ADS" - Default: "ADS"}
}
\value{
\code{list}
}
\description{
What it does
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
