% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCrossTabDS.R
\name{GetCrossTabDS}
\alias{GetCrossTabDS}
\title{GetCrossTabDS}
\usage{
GetCrossTabDS(TableName.S, FeatureNames.S, RemoveNA.S = FALSE)
}
\arguments{
\item{TableName.S}{\code{string} - Name of \code{data.frame}}

\item{FeatureNames.S}{\code{string} - Names of features separated by ','}

\item{RemoveNA.S}{\code{logical} - Indicating whether missing values should be removed prior to cross tabulation - Default: \code{FALSE}}
}
\value{
A \code{list} containing
\itemize{ \item CrossTab (\code{data.frame})
\item ChiSq.PValue (\code{numeric}) }
}
\description{
Perform cross tabulation with arbitrary number of table features.
}
\details{
Server-side AGGREGATE method
}
\author{
Bastian Reiter
}
