% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSampleStatisticsDS.R
\name{GetSampleStatisticsDS}
\alias{GetSampleStatisticsDS}
\title{GetSampleStatisticsDS}
\usage{
GetSampleStatisticsDS(
  TableName.S,
  MetricFeatureName.S,
  GroupingFeatureName.S = NULL,
  RemoveMissings.S = TRUE
)
}
\arguments{
\item{TableName.S}{\code{string} | Name of the Data frame that contains the feature}

\item{MetricFeatureName.S}{\code{string} | Name of feature}

\item{GroupingFeatureName.S}{\code{string} | Name of optional grouping feature}

\item{RemoveMissings.S}{\code{logical} | With numeric features; Whether NA values should be removed before calculating statistics}
}
\value{
A \code{tibble} containing parametric and non-parametric sample statistics
}
\description{
Calculate common sample statistics for a feature of a data frame.
}
\author{
Bastian Reiter
}
