% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSurvModelDS.R
\name{GetSurvModelDS}
\alias{GetSurvModelDS}
\title{GetSurvModelDS}
\usage{
GetSurvModelDS(
  TableName.S,
  TimeFeature.S,
  EventFeature.S,
  CovariateA.S = NULL,
  CovariateB.S = NULL,
  CovariateC.S = NULL,
  MinFollowUpTime.S = 1
)
}
\arguments{
\item{TableName.S}{\code{string} | Name of the data frame that holds time and event features}

\item{TimeFeature.S}{\code{string} | Name of time feature}

\item{EventFeature.S}{\code{string} | Name of event feature}

\item{CovariateA.S}{\code{string} | Name of optional Covariate A}

\item{CovariateB.S}{\code{string} | Name of optional Covariate B}

\item{CovariateC.S}{\code{string} | Name of optional Covariate C}

\item{MinFollowUpTime.S}{\code{integer} | Optional minimum of observed follow up time}
}
\value{
A \code{survival curve object}
}
\description{
Return Kaplan-Meier curve
}
