% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizeData.R
\name{HarmonizeData}
\alias{HarmonizeData}
\title{HarmonizeData}
\usage{
HarmonizeData(
  DataFrame,
  EligibleValueSets,
  Methods,
  TransformativeExpressions,
  TransformativeExpressions.Profile,
  Dictionary,
  Dictionary.Profile,
  FuzzyStringMatching,
  FuzzyStringMatching.Profile
)
}
\arguments{
\item{DataFrame}{\code{data.frame} containing data to be transformed}

\item{EligibleValueSets}{\code{list} of character vectors containing sets of eligible values for corresponding feature}

\item{Methods}{\code{data.frame} - Contains data on the selection of methods to use for each feature}

\item{TransformativeExpressions}{\code{data.frame} - Contains set of expressions like functions used to transform data values}

\item{TransformativeExpressions.Profile}{\code{string} - Profile selected in \emph{TransformativeExpressions}}

\item{Dictionary}{\code{data.frame} - Contains dictionary data used to look up and replace data values}

\item{Dictionary.Profile}{\code{string} - Profile selected in \emph{Dictionary}}

\item{FuzzyStringMatching}{\code{data.frame} - Contains settings for Fuzzy String Matching}

\item{FuzzyStringMatching.Profile}{\code{string} - Profile selected in \emph{FuzzyStringMatching}}
}
\value{
The input \code{data.frame} with transformed data values
}
\description{
This function addresses harmonizing data transformation for features of a \code{data.frame} based on passed settings / meta data.
Harmonization is attempted in a step-wise approach, incorporating various methods in a defined order:
\enumerate{\item Transformative expressions (e.g. functions like \code{str_to_upper()})
\item Dictionary look-up
\item Fuzzy String Matching}
}
\author{
Bastian Reiter
}
