% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_FeatureNames}
\alias{Meta_FeatureNames}
\title{CCP meta data: Feature names}
\format{
\subsection{\code{Meta_FeatureNames}}{

Tibble
\describe{
\item{TableName_Curated}{Table selector}
\item{FeaturePosition}{Position of Feature in Table}
\item{FeatureName_Raw}{Feature name in Raw Data Model}
\item{FeatureName_Curated}{Corresponding feature name in Curated Data Model}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_FeatureNames
}
\description{
A tibble containing meta data about corresponding feature names in Raw Data Model (RDM) and Curated Data Model (CDM)
}
\keyword{datasets}
