% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_TableNames}
\alias{Meta_TableNames}
\title{CCP Meta Data: Table names}
\format{
\subsection{\code{Meta_TableNames}}{

Tibble
\describe{
\item{TableName_Raw}{Table name in Raw Data Model}
\item{TableName_Curated}{Corresponding table name in Curated Data Model}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Meta_TableNames
}
\description{
A tibble containing meta data about corresponding table names in Raw Data Model (RDM) and Curated Data Model (CDM)
}
\keyword{datasets}
