% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Meta_ValueSets}
\alias{Meta_ValueSets}
\title{CCP Meta Data: Eligible value sets}
\format{
\subsection{\code{Meta_ValueSets}}{

Data frame
\describe{
\item{Table}{Table selector}
\item{FeatureRank}{Rough position of Feature in Table}
\item{Feature}{Feature label}
\item{ScaleLevel}{Scale level of feature}
\item{OriginalValue}{Value in original data}
\item{Value}{Value as preferred}
\item{Label}{Label for coded feature values}
\item{LabelOriginal}{Label in original data}
\item{FactorRank}{Used to determine order in values}
}
}
}
\source{
Own preferences
}
\usage{
Meta_ValueSets
}
\description{
A data frame containing meta data about the eligibility and transformation of CCP data values
}
\keyword{datasets}
