% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Set.DataRemediation}
\alias{Set.DataRemediation}
\title{Settings on how feature values should be remediated and which methods to use for each feature}
\format{
\subsection{\code{Set.DataRemediation}}{

\code{tibble}
\describe{
\item{Profile}{}
\item{FeatureID}{}
\item{Table}{}
\item{Feature}{}
\item{HasEligibleValueSet}{}
\item{RunRemediation}{}
\item{RemediationOrder}{}
\item{Method.TransformativeExpressions}{}
\item{Method.Dictionary}{}
\item{Method.FuzzyStringMatching}{}
\item{Method.NaiveBayes}{}
\item{MatchToDictionaryLookupsInFSM}{}
\item{UnremediatedValues.Substitute}{}
\item{UnremediatedValues.Substitution}{}
}
}
}
\source{
\url{https://github.com/BastianReiter/dsCCPhos/blob/main/Development/MetaData}
}
\usage{
Set.DataRemediation
}
\description{
A tibble
}
\author{
Bastian Reiter
}
\keyword{datasets}
