% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeEventData.R
\name{SummarizeEventData}
\alias{SummarizeEventData}
\title{SummarizeEventData}
\usage{
SummarizeEventData(EventData, ProgressBarObject = NULL)
}
\arguments{
\item{EventData}{\code{data.frame} that contains entries from \code{ADS$Events}}

\item{ProgressBarObject}{Optionally pass object of type \code{progress::progress_bar} to display progress}
}
\value{
\code{data.frame} containing summarized information
}
\description{
Auxiliary function within \code{\link{AugmentDataDS()}}
}
\details{
Summarize data stored in \code{ADS$Events} and create meaningful features for \code{ADS$Diagnosis} and \code{ADS$Patient}
}
\author{
Bastian Reiter
}
