% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransformData.R
\name{TransformData}
\alias{TransformData}
\title{TransformData}
\usage{
TransformData(DataFrame, TableName, RuleSet, RuleProfile)
}
\arguments{
\item{DataFrame}{data frame containing data to be transformed}

\item{TableName}{String | Name of the table (to enable mapping to transformation rules)}

\item{RuleSet}{Data frame | Contains predefined set of transformation rules}

\item{RuleProfile}{String | Profile name stated in rule set data frame}
}
\value{
The input data frame with transformed data values
}
\description{
Transforms data based on a given rule set. First, general transformations are performed using regular expression based rules. Second, for direct replacement of values hash tables are being compiled and applied.
}
\details{
Uses dplyr::mutate() to transform data based on transformation rules defined in a given rule set (Default: dsCCPhos::RuleSet_RawDataTransformation)
}
\author{
Bastian Reiter
}
